test: split write.rs into topic-focused files
