--- kernel/tests/write.rs
@@ -1,4857 +1,0 @@
-use std::collections::HashMap;
-use std::sync::Arc;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
-use delta_kernel::actions::deletion_vector::{DeletionVectorDescriptor, DeletionVectorStorageType};
-use delta_kernel::arrow::array::{
-    Array, ArrayRef, BinaryArray, Int32Array, Int64Array, ListArray, MapArray, StringArray,
-    StructArray, TimestampMicrosecondArray,
-};
-use delta_kernel::arrow::buffer::{NullBuffer, OffsetBuffer};
-use delta_kernel::arrow::datatypes::{DataType as ArrowDataType, Field, Schema as ArrowSchema};
-use delta_kernel::arrow::error::ArrowError;
-use delta_kernel::arrow::record_batch::RecordBatch;
-use delta_kernel::committer::FileSystemCommitter;
-use delta_kernel::engine::arrow_conversion::{TryFromKernel, TryIntoArrow as _};
-use delta_kernel::engine::arrow_data::ArrowEngineData;
-use delta_kernel::engine::default::executor::tokio::{
-    TokioBackgroundExecutor, TokioMultiThreadExecutor,
-};
-use delta_kernel::engine::default::parquet::DefaultParquetHandler;
-use delta_kernel::engine::default::{DefaultEngine, DefaultEngineBuilder};
-use delta_kernel::engine_data::FilteredEngineData;
-use delta_kernel::expressions::{column_expr, ColumnName, Scalar};
-use delta_kernel::object_store::local::LocalFileSystem;
-use delta_kernel::object_store::path::Path;
-use delta_kernel::object_store::{DynObjectStore, ObjectStoreExt as _};
-use delta_kernel::parquet::file::reader::{FileReader, SerializedFileReader};
-use delta_kernel::schema::{
-    ArrayType, ColumnMetadataKey, DataType, MapType, MetadataValue, SchemaRef, StructField,
-    StructType,
-};
-use delta_kernel::table_features::{get_any_level_column_physical_name, ColumnMappingMode};
-use delta_kernel::transaction::create_table::create_table as create_table_txn;
-use delta_kernel::transaction::CommitResult;
-use delta_kernel::{
-    DeltaResult, Engine, Error as KernelError, Expression as Expr, FileMeta, Predicate as Pred,
-    Snapshot, Version,
-};
-use itertools::Itertools;
-use serde_json::{json, Deserializer};
-use tempfile::{tempdir, TempDir};
-use test_utils::{
-    assert_partition_values, assert_result_error_with_message, assert_schema_has_field,
-    copy_directory, create_add_files_metadata, create_default_engine,
-    create_default_engine_mt_executor, create_table, create_table_and_load_snapshot,
-    engine_store_setup, nested_batches, nested_schema, read_actions_from_commit, read_add_infos,
-    remove_all_and_get_remove_actions, resolve_field, set_json_value, setup_test_tables, test_read,
-    test_table_setup, test_table_setup_mt, write_batch_to_table,
-};
-use url::Url;
-use uuid::Uuid;
-
-mod common;
-
-/// Returns the native parquet `field_id` for a field at the given physical path in a parquet file,
-/// or `None` if the field has no `field_id` set.
-///
-/// Panics if the file cannot be read or the physical path doesn't exist in the parquet schema.
-fn get_parquet_field_id(parquet_file: &std::path::Path, physical_path: &[String]) -> Option<i32> {
-    let file = std::fs::File::open(parquet_file).unwrap();
-    let reader = SerializedFileReader::new(file).unwrap();
-    let root = reader
-        .metadata()
-        .file_metadata()
-        .schema_descr()
-        .root_schema()
-        .clone();
-
-    let mut current = &root;
-    for name in physical_path {
-        current = current
-            .get_fields()
-            .iter()
-            .find(|f| f.name() == name)
-            .unwrap_or_else(|| panic!("parquet schema missing field '{name}'"));
-    }
-
-    let info = current.get_basic_info();
-    info.has_id().then(|| info.id())
-}
-
-fn validate_txn_id(commit_info: &serde_json::Value) {
-    let txn_id = commit_info["txnId"]
-        .as_str()
-        .expect("txnId should be present in commitInfo");
-    Uuid::parse_str(txn_id).expect("txnId should be valid UUID format");
-}
-
-fn validate_timestamp(commit_info: &serde_json::Value) {
-    let timestamp = commit_info["timestamp"]
-        .as_i64()
-        .expect("timestamp should be present in commitInfo");
-    let current_ts: i64 = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_millis()
-        .try_into()
-        .unwrap();
-    let two_days_ms = Duration::from_secs(2 * 24 * 60 * 60).as_millis() as i64;
-    assert!(
-        (timestamp <= current_ts && timestamp > current_ts - two_days_ms),
-        "commit timestamp should be at most 2 days behind current system time: got {timestamp}, now {current_ts}"
-    );
-}
-
-const ZERO_UUID: &str = "00000000-0000-0000-0000-000000000000";
-
-/// Creates a table with deletion vector support and writes the specified files
-async fn create_dv_table_with_files(
-    table_name: &str,
-    schema: Arc<StructType>,
-    file_paths: &[&str],
-) -> Result<
-    (
-        Arc<DynObjectStore>,
-        Arc<dyn delta_kernel::Engine>,
-        Url,
-        Vec<String>,
-    ),
-    Box<dyn std::error::Error>,
-> {
-    let (store, engine, table_url) = engine_store_setup(table_name, None);
-    let engine = Arc::new(engine);
-
-    // Create table with DV support (protocol 3/7 with deletionVectors feature)
-    create_table(
-        store.clone(),
-        table_url.clone(),
-        schema.clone(),
-        &[],
-        true, // use_37_protocol
-        vec!["deletionVectors"],
-        vec!["deletionVectors"],
-    )
-    .await?;
-
-    // Write files
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test engine")
-        .with_operation("WRITE".to_string())
-        .with_data_change(true);
-
-    let add_files_schema = txn.add_files_schema();
-
-    // Build metadata for all files at once
-    let files: Vec<(&str, i64, i64, i64)> = file_paths
-        .iter()
-        .enumerate()
-        .map(|(i, &path)| {
-            (
-                path,
-                1024 + i as i64 * 100, // size
-                1000000 + i as i64,    // mod_time
-                3,                     // num_records
-            )
-        })
-        .collect();
-    let metadata = create_add_files_metadata(add_files_schema, files)?;
-    txn.add_files(metadata);
-
-    let _ = txn.commit(engine.as_ref())?;
-
-    let paths: Vec<String> = file_paths.iter().map(|&s| s.to_string()).collect();
-    Ok((store, engine, table_url, paths))
-}
-
-/// Extracts scan files from a snapshot for use in deletion vector updates
-fn get_scan_files(
-    snapshot: Arc<Snapshot>,
-    engine: &dyn delta_kernel::Engine,
-) -> DeltaResult<Vec<FilteredEngineData>> {
-    let scan = snapshot.scan_builder().build()?;
-    let all_scan_metadata: Vec<_> = scan.scan_metadata(engine)?.collect::<Result<Vec<_>, _>>()?;
-
-    Ok(all_scan_metadata
-        .into_iter()
-        .map(|sm| sm.scan_files)
-        .collect())
-}
-
-fn get_simple_int_schema() -> Arc<StructType> {
-    Arc::new(StructType::try_new(vec![StructField::nullable("number", DataType::INTEGER)]).unwrap())
-}
-
-/// Write a metadata-update commit that sets a table property on the existing table.
-/// Returns a fresh snapshot reflecting the new commit.
-/// Used in tests as a hack to set table properties when create table doesn't support the property.
-fn set_table_properties(
-    table_path: &str,
-    table_url: &Url,
-    engine: &dyn Engine,
-    current_version: Version,
-    properties: &[(&str, &str)],
-) -> Result<Arc<Snapshot>, Box<dyn std::error::Error>> {
-    let v0_path = std::path::Path::new(table_path).join("_delta_log/00000000000000000000.json");
-    let mut meta: serde_json::Value = std::fs::read_to_string(&v0_path)?
-        .lines()
-        .find_map(|line| {
-            serde_json::from_str::<serde_json::Value>(line)
-                .ok()
-                .filter(|v| v.get("metaData").is_some())
-        })
-        .expect("version 0 should contain a metaData action");
-
-    for &(key, value) in properties {
-        meta["metaData"]["configuration"][key] = json!(value);
-    }
-
-    let new_commit = std::path::Path::new(table_path)
-        .join(format!("_delta_log/{:020}.json", current_version + 1));
-    std::fs::write(&new_commit, serde_json::to_string(&meta)?)?;
-    Ok(Snapshot::builder_for(table_url.clone()).build(engine)?)
-}
-
-/// Assert that the snapshot's column mapping mode matches the given `cm_mode` string,
-/// and return the resolved mode.
-fn assert_column_mapping_mode(snapshot: &Snapshot, cm_mode: &str) -> ColumnMappingMode {
-    let expected = match cm_mode {
-        "none" => ColumnMappingMode::None,
-        "name" => ColumnMappingMode::Name,
-        "id" => ColumnMappingMode::Id,
-        _ => panic!("unexpected cm_mode: {cm_mode}"),
-    };
-    let actual = snapshot
-        .table_properties()
-        .column_mapping_mode
-        .expect("column mapping mode should be set");
-    assert_eq!(actual, expected);
-    actual
-}
-
-/// Resolve a nested column inside a [`StructArray`] by walking the given field-name path,
-/// and downcast the leaf to the requested array type.
-fn resolve_struct_field<'a, T: 'static>(root: &'a StructArray, path: &[String]) -> &'a T {
-    assert!(!path.is_empty(), "path must be non-empty");
-    let mut current: &StructArray = root;
-    for (i, name) in path.iter().enumerate() {
-        let col = current
-            .column_by_name(name)
-            .unwrap_or_else(|| panic!("missing field: {name}"));
-        if i == path.len() - 1 {
-            return col
-                .as_any()
-                .downcast_ref::<T>()
-                .expect("leaf array type mismatch");
-        }
-        current = col
-            .as_any()
-            .downcast_ref::<StructArray>()
-            .unwrap_or_else(|| panic!("expected StructArray at field: {name}"));
-    }
-    unreachable!()
-}
-
-/// Navigate into a nested JSON value by following a sequence of object keys.
-/// E.g. `resolve_json_path(stats, &["address", "street"])` returns `stats["address"]["street"]`.
-fn resolve_json_path<'a>(root: &'a serde_json::Value, path: &[String]) -> &'a serde_json::Value {
-    path.iter().fold(root, |v, key| &v[key])
-}
-
-/// Assert that `stats["minValues"]` and `stats["maxValues"]` at the given physical path equal the
-/// expected values.
-fn assert_min_max_stats(
-    stats: &serde_json::Value,
-    physical_path: &[String],
-    expected_min: impl Into<serde_json::Value>,
-    expected_max: impl Into<serde_json::Value>,
-) {
-    assert_eq!(
-        *resolve_json_path(&stats["minValues"], physical_path),
-        expected_min.into()
-    );
-    assert_eq!(
-        *resolve_json_path(&stats["maxValues"], physical_path),
-        expected_max.into()
-    );
-}
-
-#[tokio::test]
-async fn test_commit_info() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema, &[], None, "test_table").await?
-    {
-        // create a transaction
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let committer = Box::new(FileSystemCommitter::new());
-        let txn = snapshot
-            .transaction(committer, &engine)?
-            .with_engine_info("default engine");
-
-        // commit!
-        let _ = txn.commit(&engine)?;
-
-        let commit1 = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commit: serde_json::Value = serde_json::from_slice(&commit1.bytes().await?)?;
-
-        validate_txn_id(&parsed_commit["commitInfo"]);
-
-        set_json_value(&mut parsed_commit, "commitInfo.timestamp", json!(0))?;
-        set_json_value(&mut parsed_commit, "commitInfo.txnId", json!(ZERO_UUID))?;
-
-        let expected_commit = json!({
-            "commitInfo": {
-                "timestamp": 0,
-                "operation": "UNKNOWN",
-                "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                "operationParameters": {},
-                "engineInfo": "default engine",
-                "txnId": ZERO_UUID,
-            }
-        });
-
-        assert_eq!(parsed_commit, expected_commit);
-    }
-    Ok(())
-}
-
-// check that the timestamps in commit_info and add actions are within 10s of SystemTime::now()
-fn check_action_timestamps<'a>(
-    parsed_commits: impl Iterator<Item = &'a serde_json::Value>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let now: i64 = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)?
-        .as_millis()
-        .try_into()
-        .unwrap();
-
-    parsed_commits.for_each(|commit| {
-        if let Some(commit_info_ts) = &commit.pointer("/commitInfo/timestamp") {
-            assert!((now - commit_info_ts.as_i64().unwrap()).abs() < 10_000);
-        }
-        if let Some(add_ts) = &commit.pointer("/add/modificationTime") {
-            assert!((now - add_ts.as_i64().unwrap()).abs() < 10_000);
-        }
-    });
-
-    Ok(())
-}
-
-// list all the files at `path` and check that all parquet files have the same size, and return
-// that size
-async fn get_and_check_all_parquet_sizes(store: Arc<DynObjectStore>, path: &str) -> u64 {
-    use futures::stream::StreamExt;
-    let files: Vec<_> = store.list(Some(&Path::from(path))).collect().await;
-    let parquet_files = files
-        .into_iter()
-        .filter(|f| match f {
-            Ok(f) => f.location.extension() == Some("parquet"),
-            Err(_) => false,
-        })
-        .collect::<Vec<_>>();
-    assert_eq!(parquet_files.len(), 2);
-    let size = parquet_files.first().unwrap().as_ref().unwrap().size;
-    assert!(parquet_files
-        .iter()
-        .all(|f| f.as_ref().unwrap().size == size));
-    size
-}
-
-async fn write_data_and_check_result_and_stats(
-    table_url: Url,
-    schema: SchemaRef,
-    engine: Arc<DefaultEngine<TokioBackgroundExecutor>>,
-    expected_since_commit: u64,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let committer = Box::new(FileSystemCommitter::new());
-    let mut txn = snapshot
-        .transaction(committer, engine.as_ref())?
-        .with_data_change(true);
-
-    // create two new arrow record batches to append
-    let append_data = [[1, 2, 3], [4, 5, 6]].map(|data| -> DeltaResult<_> {
-        let data = RecordBatch::try_new(
-            Arc::new(schema.as_ref().try_into_arrow()?),
-            vec![Arc::new(Int32Array::from(data.to_vec()))],
-        )?;
-        Ok(Box::new(ArrowEngineData::new(data)))
-    });
-
-    // write data out by spawning async tasks to simulate executors
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-    let tasks = append_data.into_iter().map(|data| {
-        // arc clones
-        let engine = engine.clone();
-        let write_context = write_context.clone();
-        tokio::task::spawn(async move {
-            engine
-                .write_parquet(data.as_ref().unwrap(), write_context.as_ref())
-                .await
-        })
-    });
-
-    let add_files_metadata = futures::future::join_all(tasks).await.into_iter().flatten();
-    for meta in add_files_metadata {
-        txn.add_files(meta?);
-    }
-
-    // commit!
-    match txn.commit(engine.as_ref())? {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(committed.commit_version(), expected_since_commit as Version);
-            assert_eq!(
-                committed.post_commit_stats().commits_since_checkpoint,
-                expected_since_commit
-            );
-            assert_eq!(
-                committed.post_commit_stats().commits_since_log_compaction,
-                expected_since_commit
-            );
-        }
-        _ => panic!("Commit should have succeeded"),
-    };
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_commit_info_action() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("default engine");
-
-        let _ = txn.commit(&engine)?;
-
-        let commit = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commits: Vec<_> = Deserializer::from_slice(&commit.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        validate_txn_id(&parsed_commits[0]["commitInfo"]);
-
-        // set timestamps to 0, paths and txn_id to known string values for comparison
-        // (otherwise timestamps are non-deterministic, paths and txn_id are random UUIDs)
-        set_json_value(&mut parsed_commits[0], "commitInfo.timestamp", json!(0))?;
-        set_json_value(&mut parsed_commits[0], "commitInfo.txnId", json!(ZERO_UUID))?;
-
-        let expected_commit = vec![json!({
-            "commitInfo": {
-                "timestamp": 0,
-                "operation": "UNKNOWN",
-                "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                "operationParameters": {},
-                "engineInfo": "default engine",
-                "txnId": ZERO_UUID
-            }
-        })];
-
-        assert_eq!(parsed_commits, expected_commit);
-    }
-    Ok(())
-}
-
-/// Verifies that when `engine_commit_info` is provided (the `Some` branch of `build_commit_info`):
-/// - The written JSON is correctly wrapped in a top-level `"commitInfo"` key.
-/// - Engine-only fields (not in `CommitInfo::to_schema()`) pass through to the log unchanged.
-/// - Fields that overlap with kernel-managed CommitInfo fields are overridden by kernel values,
-/// - All kernel-managed fields (`timestamp`, `kernelVersion`, `txnId`, `operationParameters`) are
-///   present with correct values.
-#[tokio::test]
-async fn test_commit_info_with_engine_commit_info() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema, &[], None, "test_table").await?
-    {
-        // Build engine_commit_info with:
-        //   - "myApp"    : engine-only field, must pass through unchanged.
-        //   - "myVersion": engine-only field, must pass through unchanged.
-        //   - "operation": overlapping with CommitInfo; kernel must override with "WRITE".
-        let arrow_schema = Arc::new(ArrowSchema::new(vec![
-            Field::new("myApp", ArrowDataType::Utf8, false),
-            Field::new("myVersion", ArrowDataType::Utf8, false),
-            Field::new("operation", ArrowDataType::Utf8, false),
-        ]));
-        let batch = RecordBatch::try_new(
-            arrow_schema,
-            vec![
-                Arc::new(StringArray::from(vec!["spark"])) as ArrayRef,
-                Arc::new(StringArray::from(vec!["3.5.0"])) as ArrayRef,
-                Arc::new(StringArray::from(vec!["STALE_OP"])) as ArrayRef,
-            ],
-        )?;
-        let engine_schema = Arc::new(StructType::new_unchecked(vec![
-            StructField::not_null("myApp", DataType::STRING),
-            StructField::not_null("myVersion", DataType::STRING),
-            StructField::nullable("operation", DataType::STRING),
-        ]));
-
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_operation("WRITE".to_string())
-            .with_commit_info(Box::new(ArrowEngineData::new(batch)), engine_schema);
-
-        let _ = txn.commit(&engine)?;
-
-        let commit = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commits: Vec<_> = Deserializer::from_slice(&commit.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        validate_txn_id(&parsed_commits[0]["commitInfo"]);
-        validate_timestamp(&parsed_commits[0]["commitInfo"]);
-
-        // Zero out non-deterministic fields for stable comparison.
-        set_json_value(&mut parsed_commits[0], "commitInfo.timestamp", json!(0))?;
-        set_json_value(&mut parsed_commits[0], "commitInfo.txnId", json!(ZERO_UUID))?;
-
-        // Null-valued CommitInfo fields (inCommitTimestamp, isBlindAppend, engineInfo) are
-        // omitted from the JSON — consistent with how the Delta log serializes optional fields.
-        let expected_commits = vec![json!({
-            "commitInfo": {
-                // Engine-only fields pass through unchanged.
-                "myApp": "spark",
-                "myVersion": "3.5.0",
-                // Kernel overrides the engine's stale "STALE_OP" with the real operation.
-                "operation": "WRITE",
-                // Remaining kernel-managed non-null fields are appended.
-                "operationParameters": {},
-                "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                "txnId": ZERO_UUID,
-                "timestamp": 0,
-            }
-        })];
-
-        assert_eq!(parsed_commits, expected_commits);
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        // write data out by spawning async tasks to simulate executors
-        let engine = Arc::new(engine);
-        write_data_and_check_result_and_stats(table_url.clone(), schema.clone(), engine.clone(), 1)
-            .await?;
-
-        let commit1 = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        let size =
-            get_and_check_all_parquet_sizes(store.clone(), format!("/{table_name}/").as_str())
-                .await;
-        // check that the timestamps in commit_info and add actions are within 10s of
-        // SystemTime::now() before we clear them for comparison
-        check_action_timestamps(parsed_commits.iter())?;
-        // check that the txn_id is valid before we clear it for comparison
-        validate_txn_id(&parsed_commits[0]["commitInfo"]);
-
-        // set timestamps to 0, paths and txn_id to known string values for comparison
-        // (otherwise timestamps are non-deterministic, paths and txn_id are random UUIDs)
-        set_json_value(&mut parsed_commits[0], "commitInfo.timestamp", json!(0))?;
-        set_json_value(&mut parsed_commits[0], "commitInfo.txnId", json!(ZERO_UUID))?;
-        set_json_value(&mut parsed_commits[1], "add.modificationTime", json!(0))?;
-        set_json_value(&mut parsed_commits[1], "add.path", json!("first.parquet"))?;
-        set_json_value(&mut parsed_commits[2], "add.modificationTime", json!(0))?;
-        set_json_value(&mut parsed_commits[2], "add.path", json!("second.parquet"))?;
-
-        let expected_commit = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 0,
-                    "operation": "UNKNOWN",
-                    "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                    "operationParameters": {},
-                    "txnId": ZERO_UUID
-                }
-            }),
-            json!({
-                "add": {
-                    "path": "first.parquet",
-                    "partitionValues": {},
-                    "size": size,
-                    "modificationTime": 0,
-                    "dataChange": true,
-                    "stats": "{\"numRecords\":3,\"nullCount\":{\"number\":0},\"minValues\":{\"number\":1},\"maxValues\":{\"number\":3},\"tightBounds\":true}"
-                }
-            }),
-            json!({
-                "add": {
-                    "path": "second.parquet",
-                    "partitionValues": {},
-                    "size": size,
-                    "modificationTime": 0,
-                    "dataChange": true,
-                    "stats": "{\"numRecords\":3,\"nullCount\":{\"number\":0},\"minValues\":{\"number\":4},\"maxValues\":{\"number\":6},\"tightBounds\":true}"
-                }
-            }),
-        ];
-
-        assert_eq!(parsed_commits, expected_commit);
-
-        test_read(
-            &ArrowEngineData::new(RecordBatch::try_new(
-                Arc::new(schema.as_ref().try_into_arrow()?),
-                vec![Arc::new(Int32Array::from(vec![1, 2, 3, 4, 5, 6]))],
-            )?),
-            &table_url,
-            engine,
-        )?;
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_no_add_actions() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("default engine");
-
-        // Commit without adding any add files
-        assert!(txn.commit(&engine)?.is_committed());
-
-        let commit1 = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let parsed_actions: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        // Verify that there only is a commit info action
-        assert_eq!(parsed_actions.len(), 1, "Expected only one action");
-        assert!(parsed_actions[0].get("commitInfo").is_some());
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append_twice() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, _, _) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        let engine = Arc::new(engine);
-        write_data_and_check_result_and_stats(table_url.clone(), schema.clone(), engine.clone(), 1)
-            .await?;
-        write_data_and_check_result_and_stats(table_url.clone(), schema.clone(), engine.clone(), 2)
-            .await?;
-
-        test_read(
-            &ArrowEngineData::new(RecordBatch::try_new(
-                Arc::new(schema.as_ref().try_into_arrow()?),
-                vec![Arc::new(Int32Array::from(vec![
-                    1, 2, 3, 4, 5, 6, 1, 2, 3, 4, 5, 6,
-                ]))],
-            )?),
-            &table_url,
-            engine,
-        )?;
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append_partitioned() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let partition_col = "partition";
-
-    // create a simple partitioned table: one int column named 'number', partitioned by string
-    // column named 'partition'
-    let table_schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("number", DataType::INTEGER),
-        StructField::nullable("partition", DataType::STRING),
-    ])?);
-    let data_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "number",
-        DataType::INTEGER,
-    )])?);
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(table_schema.clone(), &[partition_col], None, "test_table").await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let mut txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("default engine")
-            .with_data_change(false);
-
-        // create two new arrow record batches to append
-        let append_data = [[1, 2, 3], [4, 5, 6]].map(|data| -> DeltaResult<_> {
-            let data = RecordBatch::try_new(
-                Arc::new(data_schema.as_ref().try_into_arrow()?),
-                vec![Arc::new(Int32Array::from(data.to_vec()))],
-            )?;
-            Ok(Box::new(ArrowEngineData::new(data)))
-        });
-        let partition_vals = vec!["a", "b"];
-
-        // write data out by spawning async tasks to simulate executors
-        let engine = Arc::new(engine);
-        let tasks = append_data
-            .into_iter()
-            .zip(partition_vals)
-            .map(|(data, partition_val)| {
-                let write_context = Arc::new(
-                    txn.partitioned_write_context(HashMap::from([(
-                        partition_col.to_string(),
-                        Scalar::String(partition_val.into()),
-                    )]))
-                    .unwrap(),
-                );
-                // arc clones
-                let engine = engine.clone();
-                tokio::task::spawn(async move {
-                    engine
-                        .write_parquet(data.as_ref().unwrap(), write_context.as_ref())
-                        .await
-                })
-            });
-
-        let add_files_metadata = futures::future::join_all(tasks).await.into_iter().flatten();
-        for meta in add_files_metadata {
-            txn.add_files(meta?);
-        }
-
-        // commit!
-        assert!(txn.commit(engine.as_ref())?.is_committed());
-
-        let commit1 = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        let size =
-            get_and_check_all_parquet_sizes(store.clone(), format!("/{table_name}/").as_str())
-                .await;
-        // check that the timestamps in commit_info and add actions are within 10s of
-        // SystemTime::now() before we clear them for comparison
-        check_action_timestamps(parsed_commits.iter())?;
-        // check that the txn_id is valid before we clear it for comparison
-        validate_txn_id(&parsed_commits[0]["commitInfo"]);
-
-        // set timestamps to 0, paths and txn_id to known string values for comparison
-        // (otherwise timestamps are non-deterministic, paths and txn_id are random UUIDs)
-        set_json_value(&mut parsed_commits[0], "commitInfo.timestamp", json!(0))?;
-        set_json_value(&mut parsed_commits[0], "commitInfo.txnId", json!(ZERO_UUID))?;
-        set_json_value(&mut parsed_commits[1], "add.modificationTime", json!(0))?;
-        set_json_value(&mut parsed_commits[1], "add.path", json!("first.parquet"))?;
-        set_json_value(&mut parsed_commits[2], "add.modificationTime", json!(0))?;
-        set_json_value(&mut parsed_commits[2], "add.path", json!("second.parquet"))?;
-
-        let expected_commit = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 0,
-                    "operation": "UNKNOWN",
-                    "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                    "operationParameters": {},
-                    "engineInfo": "default engine",
-                    "txnId": ZERO_UUID
-                }
-            }),
-            json!({
-                "add": {
-                    "path": "first.parquet",
-                    "partitionValues": {
-                        "partition": "a"
-                    },
-                    "size": size,
-                    "modificationTime": 0,
-                    "dataChange": false,
-                    "stats": "{\"numRecords\":3,\"nullCount\":{\"number\":0},\"minValues\":{\"number\":1},\"maxValues\":{\"number\":3},\"tightBounds\":true}"
-                }
-            }),
-            json!({
-                "add": {
-                    "path": "second.parquet",
-                    "partitionValues": {
-                        "partition": "b"
-                    },
-                    "size": size,
-                    "modificationTime": 0,
-                    "dataChange": false,
-                    "stats": "{\"numRecords\":3,\"nullCount\":{\"number\":0},\"minValues\":{\"number\":4},\"maxValues\":{\"number\":6},\"tightBounds\":true}"
-                }
-            }),
-        ];
-
-        assert_eq!(parsed_commits, expected_commit);
-
-        test_read(
-            &ArrowEngineData::new(RecordBatch::try_new(
-                Arc::new(table_schema.as_ref().try_into_arrow()?),
-                vec![
-                    Arc::new(Int32Array::from(vec![1, 2, 3, 4, 5, 6])),
-                    Arc::new(StringArray::from(vec!["a", "a", "a", "b", "b", "b"])),
-                ],
-            )?),
-            &table_url,
-            engine,
-        )?;
-    }
-    Ok(())
-}
-
-// Verify that materialized partition columns do not get stats collected. The partition column
-// is physically present in the parquet file, but its stats should be omitted because the
-// value is already in the Add action's `partitionValues`.
-#[tokio::test]
-async fn test_materialized_partition_columns_excluded_from_stats(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let partition_col = "partition";
-    let table_schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("number", DataType::INTEGER),
-        StructField::nullable("partition", DataType::STRING),
-    ])?);
-
-    // Create a table with materializePartitionColumns writer feature
-    let (store, engine, table_location) = engine_store_setup("test_mat_part", None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        table_schema.clone(),
-        &[partition_col],
-        true, // use_37_protocol for writer features
-        vec![],
-        vec!["materializePartitionColumns"],
-    )
-    .await?;
-
-    let engine = Arc::new(engine);
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("default engine");
-
-    // With materializePartitionColumns, the data batch includes the partition column
-    let arrow_schema = Arc::new(table_schema.as_ref().try_into_arrow()?);
-    let batch = RecordBatch::try_new(
-        arrow_schema,
-        vec![
-            Arc::new(Int32Array::from(vec![1, 2, 3])),
-            Arc::new(StringArray::from(vec!["a", "a", "a"])),
-        ],
-    )?;
-    let data = Box::new(ArrowEngineData::new(batch));
-
-    let write_context = txn.partitioned_write_context(HashMap::from([(
-        partition_col.to_string(),
-        Scalar::String("a".into()),
-    )]))?;
-    let result = engine.write_parquet(&data, &write_context).await?;
-    txn.add_files(result);
-    assert!(txn.commit(engine.as_ref())?.is_committed());
-
-    // Read the commit log and verify stats
-    let commit = store
-        .get(&Path::from(
-            "/test_mat_part/_delta_log/00000000000000000001.json",
-        ))
-        .await?;
-    let parsed: Vec<serde_json::Value> = Deserializer::from_slice(&commit.bytes().await?)
-        .into_iter::<serde_json::Value>()
-        .try_collect()?;
-
-    let add = parsed
-        .iter()
-        .find(|v| v.get("add").is_some())
-        .expect("should have an add action");
-    let stats: serde_json::Value =
-        serde_json::from_str(add["add"]["stats"].as_str().unwrap()).unwrap();
-
-    // Stats should contain the data column but NOT the partition column
-    assert!(
-        stats["minValues"].get("number").is_some(),
-        "Data column 'number' should have minValues"
-    );
-    assert!(
-        stats["maxValues"].get("number").is_some(),
-        "Data column 'number' should have maxValues"
-    );
-    assert!(
-        stats["minValues"].get("partition").is_none(),
-        "Partition column should not have minValues even when materialized"
-    );
-    assert!(
-        stats["maxValues"].get("partition").is_none(),
-        "Partition column should not have maxValues even when materialized"
-    );
-    assert!(
-        stats["nullCount"].get("partition").is_none(),
-        "Partition column should not have nullCount even when materialized"
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append_invalid_schema() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    // create a simple table: one int column named 'number'
-    let table_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "number",
-        DataType::INTEGER,
-    )])?);
-    // incompatible data schema: one string column named 'string'
-    let data_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "string",
-        DataType::STRING,
-    )])?);
-
-    for (table_url, engine, _store, _table_name) in
-        setup_test_tables(table_schema, &[], None, "test_table").await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("default engine");
-
-        // create two new arrow record batches to append
-        let append_data = [["a", "b"], ["c", "d"]].map(|data| -> DeltaResult<_> {
-            let data = RecordBatch::try_new(
-                Arc::new(data_schema.as_ref().try_into_arrow()?),
-                vec![Arc::new(StringArray::from(data.to_vec()))],
-            )?;
-            Ok(Box::new(ArrowEngineData::new(data)))
-        });
-
-        // write data out by spawning async tasks to simulate executors
-        let engine = Arc::new(engine);
-        let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-        let tasks = append_data.into_iter().map(|data| {
-            // arc clones
-            let engine = engine.clone();
-            let write_context = write_context.clone();
-            tokio::task::spawn(async move {
-                engine
-                    .write_parquet(data.as_ref().unwrap(), write_context.as_ref())
-                    .await
-            })
-        });
-
-        let mut add_files_metadata = futures::future::join_all(tasks).await.into_iter().flatten();
-        assert!(add_files_metadata.all(|res| match res {
-            Err(KernelError::Arrow(ArrowError::InvalidArgumentError(_))) => true,
-            Err(KernelError::Backtraced { source, .. })
-                if matches!(
-                    &*source,
-                    KernelError::Arrow(ArrowError::InvalidArgumentError(_))
-                ) =>
-                true,
-            _ => false,
-        }));
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_write_txn_actions() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, store, table_name) in
-        setup_test_tables(schema, &[], None, "test_table").await?
-    {
-        // can't have duplicate app_id in same transaction
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        assert!(matches!(
-            snapshot
-                .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-                .with_transaction_id("app_id1".to_string(), 0)
-                .with_transaction_id("app_id1".to_string(), 1)
-                .commit(&engine),
-            Err(KernelError::Generic(msg)) if msg == "app_id app_id1 already exists in transaction"
-        ));
-
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("default engine")
-            .with_transaction_id("app_id1".to_string(), 1)
-            .with_transaction_id("app_id2".to_string(), 2);
-
-        // commit!
-        assert!(txn.commit(&engine)?.is_committed());
-
-        let snapshot = Snapshot::builder_for(table_url.clone())
-            .at_version(1)
-            .build(&engine)?;
-        assert_eq!(snapshot.get_app_id_version("app_id1", &engine)?, Some(1));
-        assert_eq!(snapshot.get_app_id_version("app_id2", &engine)?, Some(2));
-        assert_eq!(snapshot.get_app_id_version("app_id3", &engine)?, None);
-
-        let commit1 = store
-            .get(&Path::from(format!(
-                "/{table_name}/_delta_log/00000000000000000001.json"
-            )))
-            .await?;
-
-        let mut parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-            .into_iter::<serde_json::Value>()
-            .try_collect()?;
-
-        set_json_value(&mut parsed_commits[0], "commitInfo.timestamp", json!(0)).unwrap();
-
-        let time_ms: i64 = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)?
-            .as_millis()
-            .try_into()
-            .unwrap();
-
-        // check that last_updated times are identical
-        let last_updated1 = parsed_commits[1]
-            .get("txn")
-            .unwrap()
-            .get("lastUpdated")
-            .unwrap();
-        let last_updated2 = parsed_commits[2]
-            .get("txn")
-            .unwrap()
-            .get("lastUpdated")
-            .unwrap();
-        assert_eq!(last_updated1, last_updated2);
-
-        let last_updated = parsed_commits[1]
-            .get_mut("txn")
-            .unwrap()
-            .get_mut("lastUpdated")
-            .unwrap();
-        // sanity check that last_updated time is within 10s of now
-        assert!((last_updated.as_i64().unwrap() - time_ms).abs() < 10_000);
-        *last_updated = serde_json::Value::Number(1.into());
-
-        let last_updated = parsed_commits[2]
-            .get_mut("txn")
-            .unwrap()
-            .get_mut("lastUpdated")
-            .unwrap();
-        // sanity check that last_updated time is within 10s of now
-        assert!((last_updated.as_i64().unwrap() - time_ms).abs() < 10_000);
-        *last_updated = serde_json::Value::Number(2.into());
-
-        validate_txn_id(&parsed_commits[0]["commitInfo"]);
-
-        set_json_value(&mut parsed_commits[0], "commitInfo.txnId", json!(ZERO_UUID))?;
-
-        let expected_commit = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 0,
-                    "operation": "UNKNOWN",
-                    "kernelVersion": format!("v{}", env!("CARGO_PKG_VERSION")),
-                    "operationParameters": {},
-                    "engineInfo": "default engine",
-                    "txnId": ZERO_UUID
-                }
-            }),
-            json!({
-                "txn": {
-                    "appId": "app_id1",
-                    "version": 1,
-                    "lastUpdated": 1
-                }
-            }),
-            json!({
-                "txn": {
-                    "appId": "app_id2",
-                    "version": 2,
-                    "lastUpdated": 2
-                }
-            }),
-        ];
-
-        assert_eq!(parsed_commits, expected_commit);
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append_timestamp_ntz() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // create a table with TIMESTAMP_NTZ column
-    let schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "ts_ntz",
-        DataType::TIMESTAMP_NTZ,
-    )])?);
-
-    let (store, engine, table_location) = engine_store_setup("test_table_timestamp_ntz", None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec!["timestampNtz"],
-        vec!["timestampNtz"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_engine_info("default engine");
-
-    // Create Arrow data with TIMESTAMP_NTZ values including edge cases
-    // These are microseconds since Unix epoch
-    let timestamp_values = vec![
-        0i64,                  // Unix epoch (1970-01-01T00:00:00.000000)
-        1634567890123456i64,   // 2021-10-18T12:31:30.123456
-        1634567950654321i64,   // 2021-10-18T12:32:30.654321
-        1672531200000000i64,   // 2023-01-01T00:00:00.000000
-        253402300799999999i64, // 9999-12-31T23:59:59.999999 (near max valid timestamp)
-        -62135596800000000i64, // 0001-01-01T00:00:00.000000 (near min valid timestamp)
-    ];
-
-    let data = RecordBatch::try_new(
-        Arc::new(schema.as_ref().try_into_arrow()?),
-        vec![Arc::new(TimestampMicrosecondArray::from(timestamp_values))],
-    )?;
-
-    // Write data
-    let engine = Arc::new(engine);
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-
-    let add_files_metadata = engine
-        .write_parquet(&ArrowEngineData::new(data.clone()), write_context.as_ref())
-        .await?;
-
-    txn.add_files(add_files_metadata);
-
-    // Commit the transaction
-    assert!(txn.commit(engine.as_ref())?.is_committed());
-
-    // Verify the commit was written correctly
-    let commit1 = store
-        .get(&Path::from(
-            "/test_table_timestamp_ntz/_delta_log/00000000000000000001.json",
-        ))
-        .await?;
-
-    let parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-        .into_iter::<serde_json::Value>()
-        .try_collect()?;
-
-    // Check that we have the expected number of commits (commitInfo + add)
-    assert_eq!(parsed_commits.len(), 2);
-
-    // Check that the add action exists
-    assert!(parsed_commits[1].get("add").is_some());
-    // Ensure default of data change is true.
-    assert!(parsed_commits[1]
-        .get("add")
-        .unwrap()
-        .get("dataChange")
-        .unwrap()
-        .as_bool()
-        .unwrap());
-
-    // Verify the data can be read back correctly
-    test_read(&ArrowEngineData::new(data), &table_url, engine)?;
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_append_variant() -> Result<(), Box<dyn std::error::Error>> {
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    fn unshredded_variant_schema_flipped() -> DataType {
-        DataType::variant_type([
-            StructField::not_null("value", DataType::BINARY),
-            StructField::not_null("metadata", DataType::BINARY),
-        ])
-        .unwrap()
-    }
-    fn variant_arrow_type_flipped() -> ArrowDataType {
-        let metadata_field = Field::new("metadata", ArrowDataType::Binary, false);
-        let value_field = Field::new("value", ArrowDataType::Binary, false);
-        let fields = vec![value_field, metadata_field];
-        ArrowDataType::Struct(fields.into())
-    }
-
-    // create a table with VARIANT column
-    let table_schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("v", DataType::unshredded_variant()),
-        StructField::nullable("i", DataType::INTEGER),
-        StructField::nullable(
-            "nested",
-            // We flip the value and metadata fields in the actual parquet file for the test
-            StructType::try_new(vec![StructField::nullable(
-                "nested_v",
-                unshredded_variant_schema_flipped(),
-            )])?,
-        ),
-    ])?);
-
-    let write_schema = table_schema.clone();
-
-    let tmp_test_dir = tempdir()?;
-    let tmp_test_dir_url = Url::from_directory_path(tmp_test_dir.path()).unwrap();
-
-    let (store, engine, table_location) =
-        engine_store_setup("test_table_variant", Some(&tmp_test_dir_url));
-
-    // We can add shredding features as well as we are allowed to write unshredded variants
-    // into shredded tables and shredded reads are explicitly blocked in the default
-    // engine's parquet reader.
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        table_schema.clone(),
-        &[],
-        true,
-        vec!["variantType", "variantShredding-preview"],
-        vec!["variantType", "variantShredding-preview"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_data_change(true);
-
-    // First value corresponds to the variant value "1". Third value corresponds to the variant
-    // representing the JSON Object {"a":2}.
-    let metadata_v = vec![
-        Some(&[0x01, 0x00, 0x00][..]),
-        None,
-        Some(&[0x01, 0x01, 0x00, 0x01, 0x61][..]),
-    ];
-    let value_v = vec![
-        Some(&[0x0C, 0x01][..]),
-        None,
-        Some(&[0x02, 0x01, 0x00, 0x00, 0x01, 0x02][..]),
-    ];
-
-    let metadata_v_array = Arc::new(BinaryArray::from(metadata_v)) as ArrayRef;
-    let value_v_array = Arc::new(BinaryArray::from(value_v)) as ArrayRef;
-
-    // First value corresponds to the variant value "2". Third value corresponds to the variant
-    // representing the JSON Object {"b":3}.
-    let metadata_nested_v = vec![
-        Some(&[0x01, 0x00, 0x00][..]),
-        None,
-        Some(&[0x01, 0x01, 0x00, 0x01, 0x62][..]),
-    ];
-    let value_nested_v = vec![
-        Some(&[0x0C, 0x02][..]),
-        None,
-        Some(&[0x02, 0x01, 0x00, 0x00, 0x01, 0x03][..]),
-    ];
-
-    let value_nested_v_array = Arc::new(BinaryArray::from(value_nested_v)) as ArrayRef;
-    let metadata_nested_v_array = Arc::new(BinaryArray::from(metadata_nested_v)) as ArrayRef;
-
-    let variant_arrow = ArrowDataType::try_from_kernel(&DataType::unshredded_variant()).unwrap();
-    let variant_arrow_flipped = variant_arrow_type_flipped();
-
-    let i_values = vec![31, 32, 33];
-
-    let fields = match variant_arrow {
-        ArrowDataType::Struct(fields) => Ok(fields),
-        _ => Err(KernelError::Generic(
-            "Variant arrow data type is not struct.".to_string(),
-        )),
-    }?;
-    let fields_flipped = match variant_arrow_flipped {
-        ArrowDataType::Struct(fields) => Ok(fields),
-        _ => Err(KernelError::Generic(
-            "Variant arrow data type is not struct.".to_string(),
-        )),
-    }?;
-
-    let null_bitmap = NullBuffer::from_iter([true, false, true]);
-
-    let variant_v_array = StructArray::try_new(
-        fields.clone(),
-        vec![metadata_v_array, value_v_array],
-        Some(null_bitmap.clone()),
-    )?;
-
-    let variant_nested_v_array = Arc::new(StructArray::try_new(
-        fields_flipped.clone(),
-        vec![
-            value_nested_v_array.clone(),
-            metadata_nested_v_array.clone(),
-        ],
-        Some(null_bitmap.clone()),
-    )?);
-
-    let data = RecordBatch::try_new(
-        Arc::new(write_schema.as_ref().try_into_arrow()?),
-        vec![
-            // v variant
-            Arc::new(variant_v_array.clone()),
-            // i int
-            Arc::new(Int32Array::from(i_values.clone())),
-            // nested struct<nested_v variant>
-            Arc::new(StructArray::try_new(
-                vec![Field::new("nested_v", variant_arrow_type_flipped(), true)].into(),
-                vec![variant_nested_v_array.clone()],
-                None,
-            )?),
-        ],
-    )
-    .unwrap();
-
-    // Write data
-    let engine = Arc::new(engine);
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-
-    let add_files_metadata = (*engine)
-        .parquet_handler()
-        .as_any()
-        .downcast_ref::<DefaultParquetHandler<TokioBackgroundExecutor>>()
-        .unwrap()
-        .write_parquet_file(Box::new(ArrowEngineData::new(data.clone())), &write_context)
-        .await?;
-
-    txn.add_files(add_files_metadata);
-
-    // Commit the transaction
-    assert!(txn.commit(engine.as_ref())?.is_committed());
-
-    // Verify the commit was written correctly
-    let commit1_url = tmp_test_dir_url
-        .join("test_table_variant/_delta_log/00000000000000000001.json")
-        .unwrap();
-    let commit1 = store
-        .get(&Path::from_url_path(commit1_url.path()).unwrap())
-        .await?;
-
-    let parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-        .into_iter::<serde_json::Value>()
-        .try_collect()?;
-
-    // Check that we have the expected number of commits (commitInfo + add)
-    assert_eq!(parsed_commits.len(), 2);
-
-    // Check that the add action exists
-    assert!(parsed_commits[1].get("add").is_some());
-
-    // The scanned data will match the logical schema, not the physical one
-    let expected_schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("v", DataType::unshredded_variant()),
-        StructField::nullable("i", DataType::INTEGER),
-        StructField::nullable(
-            "nested",
-            StructType::try_new(vec![StructField::nullable(
-                "nested_v",
-                DataType::unshredded_variant(),
-            )])
-            .unwrap(),
-        ),
-    ])?);
-
-    // During the read, the flipped fields should be reordered into metadata, value.
-    let variant_nested_v_array_expected = Arc::new(StructArray::try_new(
-        fields,
-        vec![metadata_nested_v_array, value_nested_v_array],
-        Some(null_bitmap),
-    )?);
-    let variant_arrow_type: ArrowDataType =
-        ArrowDataType::try_from_kernel(&DataType::unshredded_variant()).unwrap();
-    let expected_data = RecordBatch::try_new(
-        Arc::new(expected_schema.as_ref().try_into_arrow()?),
-        vec![
-            // v variant
-            Arc::new(variant_v_array),
-            // i int
-            Arc::new(Int32Array::from(i_values)),
-            // nested struct<nested_v variant>
-            Arc::new(StructArray::try_new(
-                vec![Field::new("nested_v", variant_arrow_type, true)].into(),
-                vec![variant_nested_v_array_expected],
-                None,
-            )?),
-        ],
-    )
-    .unwrap();
-
-    test_read(&ArrowEngineData::new(expected_data), &table_url, engine)?;
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_shredded_variant_read_rejection() -> Result<(), Box<dyn std::error::Error>> {
-    // Ensure that shredded variants are rejected by the default engine's parquet reader
-
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-    let table_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "v",
-        DataType::unshredded_variant(),
-    )])?);
-
-    // The table will be attempted to be written in this form but be read into
-    // STRUCT<metadata: BINARY, value: BINARY>. The read should fail because the default engine
-    // currently does not support shredded reads.
-    let shredded_write_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "v",
-        DataType::try_struct_type([
-            StructField::new("metadata", DataType::BINARY, true),
-            StructField::new("value", DataType::BINARY, true),
-            StructField::new("typed_value", DataType::INTEGER, true),
-        ])?,
-    )])?);
-
-    let tmp_test_dir = tempdir()?;
-    let tmp_test_dir_url = Url::from_directory_path(tmp_test_dir.path()).unwrap();
-
-    let (store, engine, table_location) =
-        engine_store_setup("test_table_variant_2", Some(&tmp_test_dir_url));
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        table_schema.clone(),
-        &[],
-        true,
-        vec!["variantType", "variantShredding-preview"],
-        vec!["variantType", "variantShredding-preview"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_data_change(true);
-
-    // First value corresponds to the variant value "1". Third value corresponds to the variant
-    // representing the JSON Object {"a":2}.
-    let metadata_v = vec![
-        Some(&[0x01, 0x00, 0x00][..]),
-        Some(&[0x01, 0x01, 0x00, 0x01, 0x61][..]),
-    ];
-    let value_v = vec![
-        Some(&[0x0C, 0x01][..]),
-        Some(&[0x02, 0x01, 0x00, 0x00, 0x01, 0x02][..]),
-    ];
-    let typed_value_v = vec![Some(21), Some(3)];
-
-    let metadata_v_array = Arc::new(BinaryArray::from(metadata_v)) as ArrayRef;
-    let value_v_array = Arc::new(BinaryArray::from(value_v)) as ArrayRef;
-    let typed_value_v_array = Arc::new(Int32Array::from(typed_value_v)) as ArrayRef;
-
-    let variant_arrow = ArrowDataType::Struct(
-        vec![
-            Field::new("metadata", ArrowDataType::Binary, true),
-            Field::new("value", ArrowDataType::Binary, true),
-            Field::new("typed_value", ArrowDataType::Int32, true),
-        ]
-        .into(),
-    );
-
-    let fields = match variant_arrow {
-        ArrowDataType::Struct(fields) => Ok(fields),
-        _ => Err(KernelError::Generic(
-            "Variant arrow data type is not struct.".to_string(),
-        )),
-    }?;
-
-    let variant_v_array = StructArray::try_new(
-        fields.clone(),
-        vec![metadata_v_array, value_v_array, typed_value_v_array],
-        None,
-    )?;
-
-    let data = RecordBatch::try_new(
-        Arc::new(shredded_write_schema.as_ref().try_into_arrow()?),
-        vec![
-            // v variant
-            Arc::new(variant_v_array.clone()),
-        ],
-    )
-    .unwrap();
-
-    let engine = Arc::new(engine);
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-
-    let add_files_metadata = (*engine)
-        .parquet_handler()
-        .as_any()
-        .downcast_ref::<DefaultParquetHandler<TokioBackgroundExecutor>>()
-        .unwrap()
-        .write_parquet_file(Box::new(ArrowEngineData::new(data.clone())), &write_context)
-        .await?;
-
-    txn.add_files(add_files_metadata);
-
-    // Commit the transaction
-    assert!(txn.commit(engine.as_ref())?.is_committed());
-
-    // Verify the commit was written correctly
-    let commit1_url = tmp_test_dir_url
-        .join("test_table_variant_2/_delta_log/00000000000000000001.json")
-        .unwrap();
-    let commit1 = store
-        .get(&Path::from_url_path(commit1_url.path()).unwrap())
-        .await?;
-
-    let parsed_commits: Vec<_> = Deserializer::from_slice(&commit1.bytes().await?)
-        .into_iter::<serde_json::Value>()
-        .try_collect()?;
-
-    // Check that we have the expected number of commits (commitInfo + add)
-    assert_eq!(parsed_commits.len(), 2);
-
-    // Check that the add action exists
-    assert!(parsed_commits[1].get("add").is_some());
-
-    let res = test_read(&ArrowEngineData::new(data), &table_url, engine);
-    assert!(matches!(res,
-        Err(e) if e.to_string().contains("The default engine does not support shredded reads")));
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_set_domain_metadata_basic() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_basic";
-
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec!["domainMetadata"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-
-    let txn = snapshot.transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-
-    // write context does not conflict with domain metadata
-    let _write_context = txn.unpartitioned_write_context().unwrap();
-
-    // set multiple domain metadata
-    let domain1 = "app.config";
-    let config1 = r#"{"version": 1}"#;
-    let domain2 = "spark.settings";
-    let config2 = r#"{"cores": 4}"#;
-
-    assert!(txn
-        .with_domain_metadata(domain1.to_string(), config1.to_string())
-        .with_domain_metadata(domain2.to_string(), config2.to_string())
-        .commit(&engine)?
-        .is_committed());
-
-    let commit_data = store
-        .get(&Path::from(format!(
-            "/{table_name}/_delta_log/00000000000000000001.json"
-        )))
-        .await?
-        .bytes()
-        .await?;
-
-    let actions: Vec<serde_json::Value> = Deserializer::from_slice(&commit_data)
-        .into_iter()
-        .try_collect()?;
-
-    let domain_actions: Vec<_> = actions
-        .iter()
-        .filter(|v| v.get("domainMetadata").is_some())
-        .collect();
-
-    for action in &domain_actions {
-        let domain = action["domainMetadata"]["domain"].as_str().unwrap();
-        let config = action["domainMetadata"]["configuration"].as_str().unwrap();
-        assert!(!action["domainMetadata"]["removed"].as_bool().unwrap());
-
-        match domain {
-            d if d == domain1 => assert_eq!(config, config1),
-            d if d == domain2 => assert_eq!(config, config2),
-            _ => panic!("Unexpected domain: {domain}"),
-        }
-    }
-
-    let final_snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let domain1_config = final_snapshot.get_domain_metadata(domain1, &engine)?;
-    assert_eq!(domain1_config, Some(config1.to_string()));
-    let domain2_config = final_snapshot.get_domain_metadata(domain2, &engine)?;
-    assert_eq!(domain2_config, Some(config2.to_string()));
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_set_domain_metadata_errors() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_errors";
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec!["domainMetadata"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-
-    // System domain rejection
-    let txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let res = txn
-        .with_domain_metadata("delta.system".to_string(), "config".to_string())
-        .commit(&engine);
-    assert_result_error_with_message(
-        res,
-        "Cannot modify domains that start with 'delta.' as those are system controlled",
-    );
-
-    // Duplicate domain rejection
-    let txn2 = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let res = txn2
-        .with_domain_metadata("app.config".to_string(), "v1".to_string())
-        .with_domain_metadata("app.config".to_string(), "v2".to_string())
-        .commit(&engine);
-    assert_result_error_with_message(
-        res,
-        "Metadata for domain app.config already specified in this transaction",
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_set_domain_metadata_unsupported_writer_feature(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_unsupported";
-
-    // Create table WITHOUT domain metadata writer feature support
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec![],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let res = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_domain_metadata("app.config".to_string(), "test_config".to_string())
-        .commit(&engine);
-
-    assert_result_error_with_message(res, "Domain metadata operations require writer version 7 and the 'domainMetadata' writer feature");
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_remove_domain_metadata_unsupported_writer_feature(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_remove_domain_metadata_unsupported";
-
-    // Create table WITHOUT domain metadata writer feature support
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec![],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let res = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_domain_metadata_removed("app.config".to_string())
-        .commit(&engine);
-
-    assert_result_error_with_message(res, "Domain metadata operations require writer version 7 and the 'domainMetadata' writer feature");
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_remove_domain_metadata_non_existent_domain() -> Result<(), Box<dyn std::error::Error>>
-{
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_unsupported";
-
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec!["domainMetadata"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let txn = snapshot.transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-
-    let domain = "app.deprecated";
-
-    // removing domain metadata that doesn't exist should NOT write a tombstone
-    let _ = txn
-        .with_domain_metadata_removed(domain.to_string())
-        .commit(&engine)?;
-
-    let commit_data = store
-        .get(&Path::from(format!(
-            "/{table_name}/_delta_log/00000000000000000001.json"
-        )))
-        .await?
-        .bytes()
-        .await?;
-    let actions: Vec<serde_json::Value> = Deserializer::from_slice(&commit_data)
-        .into_iter()
-        .try_collect()?;
-
-    let domain_action = actions.iter().find(|v| v.get("domainMetadata").is_some());
-    assert!(
-        domain_action.is_none(),
-        "No tombstone should be written for non-existent domain"
-    );
-
-    let final_snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let config = final_snapshot.get_domain_metadata(domain, &engine)?;
-    assert_eq!(config, None);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_domain_metadata_set_remove_conflicts() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_unsupported";
-
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec!["domainMetadata"],
-    )
-    .await?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-
-    // set then remove same domain
-    let txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let err = txn
-        .with_domain_metadata("app.config".to_string(), "v1".to_string())
-        .with_domain_metadata_removed("app.config".to_string())
-        .commit(&engine)
-        .unwrap_err();
-    assert!(err
-        .to_string()
-        .contains("already specified in this transaction"));
-
-    // remove then set same domain
-    let txn2 = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let err = txn2
-        .with_domain_metadata_removed("test.domain".to_string())
-        .with_domain_metadata("test.domain".to_string(), "v1".to_string())
-        .commit(&engine)
-        .unwrap_err();
-    assert!(err
-        .to_string()
-        .contains("already specified in this transaction"));
-
-    // remove same domain twice
-    let txn3 = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let err = txn3
-        .with_domain_metadata_removed("another.domain".to_string())
-        .with_domain_metadata_removed("another.domain".to_string())
-        .commit(&engine)
-        .unwrap_err();
-    assert!(err
-        .to_string()
-        .contains("already specified in this transaction"));
-
-    // remove system domain
-    let txn4 = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let err = txn4
-        .with_domain_metadata_removed("delta.system".to_string())
-        .commit(&engine)
-        .unwrap_err();
-    assert!(err
-        .to_string()
-        .contains("Cannot modify domains that start with 'delta.' as those are system controlled"));
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_domain_metadata_set_then_remove() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let table_name = "test_domain_metadata_unsupported";
-
-    let (store, engine, table_location) = engine_store_setup(table_name, None);
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true,
-        vec![],
-        vec!["domainMetadata"],
-    )
-    .await?;
-
-    let domain = "app.config";
-    let configuration = r#"{"version": 1}"#;
-
-    // txn 1: set domain metadata
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let txn = snapshot.transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let _ = txn
-        .with_domain_metadata(domain.to_string(), configuration.to_string())
-        .commit(&engine)?;
-
-    // txn 2: remove the same domain metadata
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let txn = snapshot.transaction(Box::new(FileSystemCommitter::new()), &engine)?;
-    let _ = txn
-        .with_domain_metadata_removed(domain.to_string())
-        .commit(&engine)?;
-
-    // verify removal commit preserves the previous configuration
-    let commit_data = store
-        .get(&Path::from(format!(
-            "/{table_name}/_delta_log/00000000000000000002.json"
-        )))
-        .await?
-        .bytes()
-        .await?;
-    let actions: Vec<serde_json::Value> = Deserializer::from_slice(&commit_data)
-        .into_iter()
-        .try_collect()?;
-
-    let domain_action = actions
-        .iter()
-        .find(|v| v.get("domainMetadata").is_some())
-        .unwrap();
-    assert_eq!(domain_action["domainMetadata"]["domain"], domain);
-    assert_eq!(
-        domain_action["domainMetadata"]["configuration"],
-        configuration
-    );
-    assert_eq!(domain_action["domainMetadata"]["removed"], true);
-
-    // verify reads see the metadata removal
-    let final_snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    let domain_config = final_snapshot.get_domain_metadata(domain, &engine)?;
-    assert_eq!(domain_config, None);
-
-    Ok(())
-}
-
-async fn get_ict_at_version(
-    store: Arc<DynObjectStore>,
-    table_url: &Url,
-    version: u64,
-) -> Result<i64, Box<dyn std::error::Error>> {
-    let commit_path = table_url.join(&format!("_delta_log/{version:020}.json"))?;
-    let commit = store.get(&Path::from_url_path(commit_path.path())?).await?;
-    let commit_content = String::from_utf8(commit.bytes().await?.to_vec())?;
-
-    // Parse each line of the commit log (NDJSON format)
-    // CommitInfo MUST be the first action when ICT is enabled
-    let lines: Vec<_> = commit_content
-        .lines()
-        .filter(|line| !line.trim().is_empty())
-        .collect();
-    assert!(
-        !lines.is_empty(),
-        "Commit log at version {version} should not be empty"
-    );
-
-    // First line should contain commitInfo with inCommitTimestamp
-    let first_action: serde_json::Value = serde_json::from_str(lines[0])?;
-    let commit_info = first_action
-        .get("commitInfo")
-        .expect("First action must be commitInfo when ICT is enabled");
-    let ict = commit_info
-        .get("inCommitTimestamp")
-        .expect("commitInfo must have inCommitTimestamp when ICT is enabled")
-        .as_i64()
-        .unwrap();
-    Ok(ict)
-}
-
-/// Helper function to generate a simple data file and add it to the transaction
-/// This simplifies repetitive data generation in tests
-async fn generate_and_add_data_file(
-    txn: &mut delta_kernel::transaction::Transaction,
-    engine: &DefaultEngine<TokioBackgroundExecutor>,
-    schema: SchemaRef,
-    values: Vec<i32>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let data = RecordBatch::try_new(
-        Arc::new(schema.as_ref().try_into_arrow()?),
-        vec![Arc::new(Int32Array::from(values))],
-    )?;
-
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-    let file_meta = engine
-        .write_parquet(&ArrowEngineData::new(data), write_context.as_ref())
-        .await?;
-    txn.add_files(file_meta);
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_ict_commit_e2e() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // create a simple table: one int column named 'number' with ICT enabled
-    let schema = get_simple_int_schema();
-
-    let tmp_dir = TempDir::new()?;
-    let tmp_test_dir_url = Url::from_file_path(&tmp_dir).unwrap();
-
-    let (store, engine, table_location) =
-        engine_store_setup("test_ict_first_commit", Some(&tmp_test_dir_url));
-
-    // Create table with ICT enabled (writer version 7)
-    let table_url = test_utils::create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],                       // no partition columns
-        true,                      // use protocol 3.7
-        vec![],                    // no reader features
-        vec!["inCommitTimestamp"], // Enable ICT! Note: table feature is also set.
-    )
-    .await?;
-
-    // FIRST COMMIT: This exercises version() == 0 branch and generates ICT
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    assert_eq!(
-        snapshot.version(),
-        0,
-        "Initial snapshot should be version 0"
-    );
-
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_engine_info("ict test");
-
-    // Add some data
-    generate_and_add_data_file(&mut txn, &engine, schema.clone(), vec![1, 2, 3]).await?;
-
-    // First commit
-    let commit_result = txn.commit(&engine)?;
-    match commit_result {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(
-                committed.commit_version(),
-                1,
-                "First commit should result in version 1"
-            );
-        }
-        CommitResult::ConflictedTransaction(conflicted) => {
-            panic!(
-                "First commit should not conflict, got conflict at version {}",
-                conflicted.conflict_version()
-            );
-        }
-        CommitResult::RetryableTransaction(_) => {
-            panic!("First commit should not be retryable error");
-        }
-    }
-
-    // VERIFY: Check that the commit log contains inCommitTimestamp
-    let first_ict = get_ict_at_version(store.clone(), &table_url, 1).await?;
-
-    assert!(
-        first_ict > 1612345678,
-        "First commit ICT ({first_ict}) should be greater than enablement timestamp (1612345678)"
-    );
-
-    // Second commit
-    let snapshot2 = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-    assert_eq!(
-        snapshot2.version(),
-        1,
-        "Second snapshot should be version 1"
-    );
-
-    let mut txn2 = snapshot2
-        .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-        .with_engine_info("ict test 2");
-
-    // Add more data
-    generate_and_add_data_file(&mut txn2, &engine, schema, vec![4, 5, 6]).await?;
-
-    // Second commit
-    let commit_result2 = txn2.commit(&engine)?;
-    match commit_result2 {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(
-                committed.commit_version(),
-                2,
-                "Second commit should result in version 2"
-            );
-        }
-        CommitResult::ConflictedTransaction(conflicted) => {
-            panic!(
-                "Second commit should not conflict, got conflict at version {}",
-                conflicted.conflict_version()
-            );
-        }
-        CommitResult::RetryableTransaction(_) => {
-            panic!("Second commit should not be retryable error");
-        }
-    }
-
-    // VERIFY: Check that second commit has proper monotonic ICT
-    let second_ict = get_ict_at_version(store, &table_url, 2).await?;
-
-    // Verify monotonic property: second_ict > first_ict
-    assert!(
-        second_ict > first_ict,
-        "Second ICT ({second_ict}) should be greater than first ICT ({first_ict})"
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_remove_files_adds_expected_entries() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that Remove actions generated from scan metadata contain all expected
-    // fields from the Remove struct (defined in kernel/src/actions/mod.rs).
-    //
-    // This test uses the table-with-dv-small dataset which contains files with tags and deletion
-    // vectors.
-    //
-    // Not populated in the dataset are (covered by row_tracking tests):
-    // baseRowId (optional i64)
-    // defaultRowCommitVersion (optional i64)
-    use std::path::PathBuf;
-
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let tmp_dir = tempdir()?;
-    let tmp_table_path = tmp_dir.path().join("table-with-dv-small");
-    let source_path = std::fs::canonicalize(PathBuf::from("./tests/data/table-with-dv-small/"))?;
-    copy_directory(&source_path, &tmp_table_path)?;
-
-    let table_url = url::Url::from_directory_path(&tmp_table_path).unwrap();
-    let engine = create_default_engine(&table_url)?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone())
-        .at_version(1)
-        .build(engine.as_ref())?;
-
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test engine")
-        .with_data_change(true);
-
-    let scan = snapshot.scan_builder().build()?;
-    let scan_metadata = scan.scan_metadata(engine.as_ref())?.next().unwrap()?;
-
-    let (data, selection_vector) = scan_metadata.scan_files.into_parts();
-    let remove_metadata = FilteredEngineData::try_new(data, selection_vector)?;
-
-    txn.remove_files(remove_metadata);
-
-    let result = txn.commit(engine.as_ref())?;
-
-    match result {
-        CommitResult::CommittedTransaction(committed) => {
-            let commit_version = committed.commit_version();
-
-            // Read the commit log directly to verify remove actions
-            let commit_path = tmp_table_path.join(format!("_delta_log/{commit_version:020}.json"));
-            let commit_content = std::fs::read_to_string(commit_path)?;
-
-            let parsed_commits: Vec<_> = Deserializer::from_str(&commit_content)
-                .into_iter::<serde_json::Value>()
-                .try_collect()?;
-
-            // Verify we have at least commitInfo and remove actions
-            assert!(
-                parsed_commits.len() >= 2,
-                "Expected at least 2 actions (commitInfo + remove)"
-            );
-
-            // Extract the commitInfo timestamp to validate against deletionTimestamp
-            let commit_info_action = parsed_commits
-                .iter()
-                .find(|action| action.get("commitInfo").is_some())
-                .expect("Missing commitInfo action");
-            let commit_info = &commit_info_action["commitInfo"];
-            let commit_timestamp = commit_info["timestamp"]
-                .as_i64()
-                .expect("Missing timestamp in commitInfo");
-
-            // Verify remove actions
-            let remove_actions: Vec<_> = parsed_commits
-                .iter()
-                .filter(|action| action.get("remove").is_some())
-                .collect();
-
-            assert!(
-                !remove_actions.is_empty(),
-                "Expected at least one remove action"
-            );
-
-            assert_eq!(remove_actions.len(), 1);
-            let remove_action = remove_actions[0];
-            let remove = &remove_action["remove"];
-
-            // path (required)
-            assert!(remove.get("path").is_some(), "Missing path field");
-            let path = remove["path"].as_str().expect("path should be a string");
-            assert_eq!(
-                path,
-                "part-00000-fae5310a-a37d-4e51-827b-c3d5516560ca-c000.snappy.parquet"
-            );
-
-            // dataChange (required)
-            assert_eq!(remove["dataChange"].as_bool(), Some(true));
-
-            // deletionTimestamp (optional) - should match commit timestamp
-            let deletion_timestamp = remove["deletionTimestamp"]
-                .as_i64()
-                .expect("Missing deletionTimestamp");
-            assert_eq!(
-                deletion_timestamp, commit_timestamp,
-                "deletionTimestamp should match commit timestamp"
-            );
-
-            // extendedFileMetadata (optional)
-            assert_eq!(remove["extendedFileMetadata"].as_bool(), Some(true));
-
-            // partitionValues (optional)
-            let partition_vals = remove["partitionValues"]
-                .as_object()
-                .expect("Missing partitionValues");
-            assert_eq!(partition_vals.len(), 0);
-
-            // size (optional)
-            let size = remove["size"].as_i64().expect("Missing size");
-            assert_eq!(size, 635);
-
-            // stats (optional)
-            let stats = remove["stats"].as_str().expect("Missing stats");
-            let stats_json: serde_json::Value = serde_json::from_str(stats)?;
-            assert_eq!(stats_json["numRecords"], 10);
-
-            // tags (optional)
-            let tags = remove["tags"].as_object().expect("Missing tags");
-            assert_eq!(
-                tags.get("INSERTION_TIME").and_then(|v| v.as_str()),
-                Some("1677811178336000")
-            );
-            assert_eq!(
-                tags.get("MIN_INSERTION_TIME").and_then(|v| v.as_str()),
-                Some("1677811178336000")
-            );
-            assert_eq!(
-                tags.get("MAX_INSERTION_TIME").and_then(|v| v.as_str()),
-                Some("1677811178336000")
-            );
-            assert_eq!(
-                tags.get("OPTIMIZE_TARGET_SIZE").and_then(|v| v.as_str()),
-                Some("268435456")
-            );
-
-            // deletionVector (optional)
-            let dv = remove["deletionVector"]
-                .as_object()
-                .expect("Missing deletionVector");
-            assert_eq!(dv.get("storageType").and_then(|v| v.as_str()), Some("u"));
-            assert_eq!(
-                dv.get("pathOrInlineDv").and_then(|v| v.as_str()),
-                Some("vBn[lx{q8@P<9BNH/isA")
-            );
-            assert_eq!(dv.get("offset").and_then(|v| v.as_i64()), Some(1));
-            assert_eq!(dv.get("sizeInBytes").and_then(|v| v.as_i64()), Some(36));
-            assert_eq!(dv.get("cardinality").and_then(|v| v.as_i64()), Some(2));
-
-            // Row tracking fields should be absent as the feature is was not enabled on writing
-            // row_tracking tests cover having these populated.
-            assert!(remove.get("baseRowId").is_none());
-            assert!(remove.get("defaultRowCommitVersion").is_none());
-        }
-        _ => panic!("Transaction should be committed"),
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_deletion_vectors_adds_expected_entries(
-) -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that deletion vector updates write proper Remove and Add actions
-    // to the transaction log.
-    //
-    // NOTE: Additional unit tests for update_deletion_vectors exist in
-    // kernel/src/transaction/mod.rs
-    //
-    // The test validates:
-    // 1. Transaction setup for DV updates
-    // 2. Scanning and extracting scan files with DV data
-    // 3. Creating new DV descriptors for the files
-    // 4. Calling update_deletion_vectors to update the DVs
-    // 5. Committing and verifying the generated actions
-    //
-    // Expected commit log structure:
-    // - commitInfo: Contains metadata about the transaction
-    // - remove: Contains OLD deletion vector data and original file metadata
-    // - add: Contains NEW deletion vector data and updated file metadata
-    //
-    // The test ensures:
-    // - Remove action has the OLD DV descriptor with all 5 fields
-    // - Add action has the NEW DV descriptor with all 5 fields
-    // - All file metadata is preserved (size, stats, tags, partitionValues)
-    // - dataChange is properly set to true
-    // - deletionTimestamp matches commit timestamp
-    use std::path::PathBuf;
-
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let tmp_dir = tempdir()?;
-    let tmp_table_path = tmp_dir.path().join("table-with-dv-small");
-    let source_path = std::fs::canonicalize(PathBuf::from("./tests/data/table-with-dv-small/"))?;
-    copy_directory(&source_path, &tmp_table_path)?;
-
-    let table_url = url::Url::from_directory_path(&tmp_table_path).unwrap();
-    let engine = create_default_engine(&table_url)?;
-
-    let snapshot = Snapshot::builder_for(table_url.clone())
-        .at_version(1)
-        .build(engine.as_ref())?;
-
-    // Create transaction with DV update mode enabled
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test engine")
-        .with_operation("UPDATE".to_string())
-        .with_data_change(true);
-
-    // Build scan and collect all scan metadata
-    let scan = snapshot.clone().scan_builder().build()?;
-    let all_scan_metadata: Vec<_> = scan
-        .scan_metadata(engine.as_ref())?
-        .collect::<Result<Vec<_>, _>>()?;
-
-    // Extract scan files for DV update
-    let scan_files: Vec<_> = all_scan_metadata
-        .into_iter()
-        .map(|sm| sm.scan_files)
-        .collect();
-
-    // Create new DV descriptors for the files
-    let file_path = "part-00000-fae5310a-a37d-4e51-827b-c3d5516560ca-c000.snappy.parquet";
-    let mut dv_map = HashMap::new();
-
-    // Create a NEW deletion vector descriptor (different from the original)
-    let new_dv = DeletionVectorDescriptor {
-        storage_type: DeletionVectorStorageType::PersistedRelative,
-        path_or_inline_dv: "cd^-aqEH.-t@S}K{vb[*k^".to_string(),
-        offset: Some(10),
-        size_in_bytes: 40,
-        cardinality: 3,
-    };
-    dv_map.insert(file_path.to_string(), new_dv);
-
-    // Call update_deletion_vectors to exercise the API
-    txn.update_deletion_vectors(dv_map, scan_files.into_iter().map(Ok))?;
-
-    // Commit the transaction
-    let result = txn.commit(engine.as_ref())?;
-
-    match result {
-        CommitResult::CommittedTransaction(committed) => {
-            let commit_version = committed.commit_version();
-
-            // Read the original version 1 log to get original file metadata
-            let original_log_path = tmp_table_path.join("_delta_log/00000000000000000001.json");
-            let original_log_content = std::fs::read_to_string(original_log_path)?;
-            let original_commits: Vec<_> = Deserializer::from_str(&original_log_content)
-                .into_iter::<serde_json::Value>()
-                .try_collect()?;
-
-            let file_path = "part-00000-fae5310a-a37d-4e51-827b-c3d5516560ca-c000.snappy.parquet";
-
-            // Extract original file metadata from version 1
-            let original_add = original_commits
-                .iter()
-                .find(|action| {
-                    action
-                        .get("add")
-                        .and_then(|add| add.get("path").and_then(|p| p.as_str()))
-                        == Some(file_path)
-                })
-                .expect("Missing original add action in version 1")
-                .get("add")
-                .expect("Should have add field");
-
-            let original_size = original_add["size"]
-                .as_i64()
-                .expect("Original add action should have size");
-            let original_partition_values = original_add["partitionValues"]
-                .as_object()
-                .expect("Original add action should have partitionValues");
-            let original_tags = original_add.get("tags");
-            let original_stats = original_add.get("stats");
-
-            // Read the commit log directly
-            let commit_path = tmp_table_path.join(format!("_delta_log/{commit_version:020}.json"));
-            let commit_content = std::fs::read_to_string(commit_path)?;
-
-            let parsed_commits: Vec<_> = Deserializer::from_str(&commit_content)
-                .into_iter::<serde_json::Value>()
-                .try_collect()?;
-
-            // Should have commitInfo, remove, and add actions
-            assert!(
-                parsed_commits.len() >= 3,
-                "Expected at least 3 actions (commitInfo + remove + add), got {}",
-                parsed_commits.len()
-            );
-
-            // Extract commitInfo timestamp
-            let commit_info_action = parsed_commits
-                .iter()
-                .find(|action| action.get("commitInfo").is_some())
-                .expect("Missing commitInfo action");
-            let commit_info = &commit_info_action["commitInfo"];
-            let commit_timestamp = commit_info["timestamp"]
-                .as_i64()
-                .expect("Missing timestamp in commitInfo");
-
-            // Verify remove action contains OLD DV information
-            let remove_actions: Vec<_> = parsed_commits
-                .iter()
-                .filter(|action| action.get("remove").is_some())
-                .collect();
-
-            assert_eq!(
-                remove_actions.len(),
-                1,
-                "Expected exactly one remove action"
-            );
-
-            let remove_action = remove_actions[0];
-            let remove = &remove_action["remove"];
-
-            assert_eq!(
-                remove["path"].as_str(),
-                Some(file_path),
-                "Remove path should match"
-            );
-            assert_eq!(remove["dataChange"].as_bool(), Some(true));
-            assert_eq!(
-                remove["deletionTimestamp"].as_i64(),
-                Some(commit_timestamp),
-                "deletionTimestamp should match commit timestamp"
-            );
-
-            // Verify OLD deletion vector in remove action
-            let old_dv = remove["deletionVector"]
-                .as_object()
-                .expect("Remove action should have deletionVector");
-            assert_eq!(
-                old_dv.get("storageType").and_then(|v| v.as_str()),
-                Some("u"),
-                "Old DV storage type should be 'u'"
-            );
-            assert_eq!(
-                old_dv.get("pathOrInlineDv").and_then(|v| v.as_str()),
-                Some("vBn[lx{q8@P<9BNH/isA"),
-                "Old DV path should match original"
-            );
-            assert_eq!(
-                old_dv.get("offset").and_then(|v| v.as_i64()),
-                Some(1),
-                "Old DV offset should be 1"
-            );
-            assert_eq!(
-                old_dv.get("sizeInBytes").and_then(|v| v.as_i64()),
-                Some(36),
-                "Old DV size should be 36"
-            );
-            assert_eq!(
-                old_dv.get("cardinality").and_then(|v| v.as_i64()),
-                Some(2),
-                "Old DV cardinality should be 2"
-            );
-
-            // Verify file metadata is preserved in remove action
-            let remove_size = remove["size"]
-                .as_i64()
-                .expect("Remove action should have size");
-            let remove_partition_values = remove["partitionValues"]
-                .as_object()
-                .expect("Remove action should have partitionValues");
-            let remove_tags = remove.get("tags");
-            let remove_stats = remove.get("stats");
-
-            // Verify add action contains NEW DV information
-            let add_actions: Vec<_> = parsed_commits
-                .iter()
-                .filter(|action| action.get("add").is_some())
-                .collect();
-
-            assert_eq!(add_actions.len(), 1, "Expected exactly one add action");
-
-            let add_action = add_actions[0];
-            let add = &add_action["add"];
-
-            assert_eq!(
-                add["path"].as_str(),
-                Some(file_path),
-                "Add path should match"
-            );
-            assert_eq!(add["dataChange"].as_bool(), Some(true));
-
-            // Verify NEW deletion vector in add action
-            let new_dv = add["deletionVector"]
-                .as_object()
-                .expect("Add action should have deletionVector");
-            assert_eq!(
-                new_dv.get("storageType").and_then(|v| v.as_str()),
-                Some("u"),
-                "New DV storage type should be 'u'"
-            );
-            assert_eq!(
-                new_dv.get("pathOrInlineDv").and_then(|v| v.as_str()),
-                Some("cd^-aqEH.-t@S}K{vb[*k^"),
-                "New DV path should match updated value"
-            );
-            assert_eq!(
-                new_dv.get("offset").and_then(|v| v.as_i64()),
-                Some(10),
-                "New DV offset should be 10"
-            );
-            assert_eq!(
-                new_dv.get("sizeInBytes").and_then(|v| v.as_i64()),
-                Some(40),
-                "New DV size should be 40"
-            );
-            assert_eq!(
-                new_dv.get("cardinality").and_then(|v| v.as_i64()),
-                Some(3),
-                "New DV cardinality should be 3"
-            );
-
-            // Verify file metadata is preserved in add action
-            let add_size = add["size"].as_i64().expect("Add action should have size");
-            let add_partition_values = add["partitionValues"]
-                .as_object()
-                .expect("Add action should have partitionValues");
-            let add_tags = add.get("tags");
-            let add_stats = add.get("stats");
-
-            // Ensure metadata is consistent between remove and add actions
-            assert_eq!(
-                remove_size, add_size,
-                "File size should be preserved between remove and add"
-            );
-            assert_eq!(
-                remove_partition_values, add_partition_values,
-                "Partition values should be preserved between remove and add"
-            );
-            assert_eq!(
-                remove_tags, add_tags,
-                "Tags should be preserved between remove and add"
-            );
-            assert_eq!(
-                remove_stats, add_stats,
-                "Stats should be preserved between remove and add"
-            );
-
-            // Ensure metadata matches the original file metadata from version 1
-            assert_eq!(
-                remove_size, original_size,
-                "Remove action size should match original file size"
-            );
-            assert_eq!(
-                add_size, original_size,
-                "Add action size should match original file size"
-            );
-            assert_eq!(
-                remove_partition_values, original_partition_values,
-                "Remove action partition values should match original"
-            );
-            assert_eq!(
-                add_partition_values, original_partition_values,
-                "Add action partition values should match original"
-            );
-            assert_eq!(
-                remove_tags, original_tags,
-                "Remove action tags should match original"
-            );
-            assert_eq!(
-                add_tags, original_tags,
-                "Add action tags should match original"
-            );
-            assert_eq!(
-                remove_stats, original_stats,
-                "Remove action stats should match original"
-            );
-            assert_eq!(
-                add_stats, original_stats,
-                "Add action stats should match original"
-            );
-        }
-        _ => panic!("Transaction should be committed"),
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_deletion_vectors_multiple_files() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that update_deletion_vectors can update multiple files
-    // in a single call, creating proper Remove and Add actions for each file.
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("id", DataType::INTEGER),
-        StructField::nullable("value", DataType::STRING),
-    ])?);
-
-    // Setup: Create table with 3 files
-    let file_names = &["file0.parquet", "file1.parquet", "file2.parquet"];
-    let (store, engine, table_url, file_paths) =
-        create_dv_table_with_files("test_table", schema, file_names).await?;
-
-    // Create DV update transaction
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test engine")
-        .with_operation("UPDATE".to_string())
-        .with_data_change(true);
-
-    let mut scan_files = get_scan_files(snapshot.clone(), engine.as_ref())?;
-
-    // Update deletion vectors for all 3 files in a single call
-    let mut dv_map = HashMap::new();
-    for (idx, file_path) in file_paths.iter().enumerate() {
-        let descriptor = DeletionVectorDescriptor {
-            storage_type: DeletionVectorStorageType::PersistedRelative,
-            path_or_inline_dv: format!("dv_file_{idx}.bin"),
-            offset: Some(idx as i32 * 10),
-            size_in_bytes: 40 + idx as i32,
-            cardinality: idx as i64 + 1,
-        };
-        dv_map.insert(file_path.to_string(), descriptor);
-    }
-
-    txn.update_deletion_vectors(dv_map, scan_files.drain(..).map(Ok))?;
-
-    // Commit the transaction
-    let result = txn.commit(engine.as_ref())?;
-
-    match result {
-        CommitResult::CommittedTransaction(committed) => {
-            let commit_version = committed.commit_version();
-
-            // Read the commit log directly from object store
-            let final_commit_path =
-                table_url.join(&format!("_delta_log/{commit_version:020}.json"))?;
-            let commit_content = store
-                .get(&Path::from_url_path(final_commit_path.path())?)
-                .await?
-                .bytes()
-                .await?;
-
-            let parsed_commits: Vec<_> = Deserializer::from_slice(&commit_content)
-                .into_iter::<serde_json::Value>()
-                .try_collect()?;
-
-            // Extract all remove and add actions
-            let remove_actions: Vec<_> = parsed_commits
-                .iter()
-                .filter(|action| action.get("remove").is_some())
-                .collect();
-
-            let add_actions: Vec<_> = parsed_commits
-                .iter()
-                .filter(|action| action.get("add").is_some())
-                .collect();
-
-            // Should have 3 remove and 3 add actions
-            assert_eq!(
-                remove_actions.len(),
-                3,
-                "Expected 3 remove actions for 3 files"
-            );
-            assert_eq!(add_actions.len(), 3, "Expected 3 add actions for 3 files");
-
-            // Verify each file has a DV in both remove and add
-            for (idx, file_path) in file_paths.iter().enumerate() {
-                // Find the remove action for this file
-                let remove_action = remove_actions
-                    .iter()
-                    .find(|action| action["remove"]["path"].as_str() == Some(file_path.as_str()))
-                    .unwrap_or_else(|| panic!("Should find remove action for {file_path}"));
-
-                // Find the add action for this file
-                let add_action = add_actions
-                    .iter()
-                    .find(|action| action["add"]["path"].as_str() == Some(file_path.as_str()))
-                    .unwrap_or_else(|| panic!("Should find add action for {file_path}"));
-
-                // Verify remove action does NOT have a DV (since these were newly written files)
-                assert!(
-                    remove_action["remove"]["deletionVector"].is_null(),
-                    "Remove action for newly written file should not have a DV"
-                );
-
-                // Verify add action has the NEW DV
-                let add_dv = add_action["add"]["deletionVector"]
-                    .as_object()
-                    .expect("Add action should have deletionVector");
-
-                let expected_path = format!("dv_file_{idx}.bin");
-                assert_eq!(
-                    add_dv.get("pathOrInlineDv").and_then(|v| v.as_str()),
-                    Some(expected_path.as_str()),
-                    "DV path should match for file {file_path}"
-                );
-                assert_eq!(
-                    add_dv.get("offset").and_then(|v| v.as_i64()),
-                    Some(idx as i64 * 10),
-                    "DV offset should match for file {file_path}"
-                );
-                assert_eq!(
-                    add_dv.get("sizeInBytes").and_then(|v| v.as_i64()),
-                    Some(40 + idx as i64),
-                    "DV size should match for file {file_path}"
-                );
-                assert_eq!(
-                    add_dv.get("cardinality").and_then(|v| v.as_i64()),
-                    Some(idx as i64 + 1),
-                    "DV cardinality should match for file {file_path}"
-                );
-            }
-        }
-        _ => panic!("Transaction should be committed"),
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_remove_files_verify_files_excluded_from_scan(
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Adds and then removes files and then verifies they don't appear in the scan.
-
-    // setup tracing
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // create a simple table: one int column named 'number'
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, _store, _table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        // First, add some files to the table
-        let engine = Arc::new(engine);
-        write_data_and_check_result_and_stats(table_url.clone(), schema.clone(), engine.clone(), 1)
-            .await?;
-
-        // Get initial file count
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-        let scan = snapshot.clone().scan_builder().build()?;
-        let scan_metadata = scan.scan_metadata(engine.as_ref())?.next().unwrap()?;
-        let (_, selection_vector) = scan_metadata.scan_files.into_parts();
-        let initial_file_count = selection_vector.iter().filter(|&x| *x).count();
-
-        assert!(initial_file_count > 0);
-
-        // Now create a transaction to remove files
-        let mut txn = snapshot
-            .clone()
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?;
-
-        // Create a new scan to get file metadata for removal
-        let scan2 = snapshot.scan_builder().build()?;
-        let scan_metadata2 = scan2.scan_metadata(engine.as_ref())?.next().unwrap()?;
-
-        // Create FilteredEngineData for removal (select all rows for removal)
-        let file_remove_count = (scan_metadata2.scan_files.data().len()
-            - scan_metadata2.scan_files.selection_vector().len())
-            + scan_metadata2
-                .scan_files
-                .selection_vector()
-                .iter()
-                .filter(|&x| *x)
-                .count();
-        assert!(file_remove_count > 0);
-
-        // Add remove files to transaction
-        txn.remove_files(scan_metadata2.scan_files);
-
-        // Commit the transaction
-        let result = txn.commit(engine.as_ref());
-
-        match result? {
-            CommitResult::CommittedTransaction(committed) => {
-                assert_eq!(committed.commit_version(), 2);
-
-                let new_snapshot = Snapshot::builder_for(table_url.clone())
-                    .at_version(2)
-                    .build(engine.as_ref())?;
-
-                let new_scan = new_snapshot.scan_builder().build()?;
-                let mut new_file_count = 0;
-                for new_metadata in new_scan.scan_metadata(engine.as_ref())? {
-                    new_file_count += new_metadata?.scan_files.data().len();
-                }
-
-                // All files were removed, so new_file_count should be zero
-                assert_eq!(new_file_count, 0);
-            }
-            _ => panic!("Transaction did not succeeed."),
-        }
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_remove_files_with_modified_selection_vector() -> Result<(), Box<dyn std::error::Error>>
-{
-    // This test verifies that we can selectively remove files by:
-    // 1. Calling remove_files multiple times with different subsets
-    // 2. Modifying the selection vector to choose which files to remove
-
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, _store, _table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_table").await?
-    {
-        let engine = Arc::new(engine);
-
-        // Write data multiple times to create multiple files
-        for i in 1..=5 {
-            write_data_and_check_result_and_stats(
-                table_url.clone(),
-                schema.clone(),
-                engine.clone(),
-                i,
-            )
-            .await?;
-        }
-
-        // Get initial file count
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-        let scan = snapshot.clone().scan_builder().build()?;
-
-        let mut initial_file_count = 0;
-        for metadata in scan.scan_metadata(engine.as_ref())? {
-            let metadata = metadata?;
-            initial_file_count += metadata
-                .scan_files
-                .selection_vector()
-                .iter()
-                .filter(|&x| *x)
-                .count();
-        }
-
-        assert!(
-            initial_file_count >= 3,
-            "Need at least 3 files for this test, got {initial_file_count}"
-        );
-
-        // Create a transaction to remove files in two batches
-        let mut txn = snapshot
-            .clone()
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-            .with_engine_info("selective remove test")
-            .with_operation("DELETE".to_string())
-            .with_data_change(true);
-
-        // First batch: Remove only the first file
-        let scan2 = snapshot.clone().scan_builder().build()?;
-        let scan_metadata2 = scan2.scan_metadata(engine.as_ref())?.next().unwrap()?;
-        let (data, mut selection_vector) = scan_metadata2.scan_files.into_parts();
-
-        // Select only the first file for removal
-        let mut first_batch_removed = 0;
-        for selected in selection_vector.iter_mut() {
-            if *selected && first_batch_removed < 1 {
-                // Keep selected for removal
-                first_batch_removed += 1;
-            } else {
-                // Don't remove
-                *selected = false;
-            }
-        }
-
-        assert_eq!(
-            first_batch_removed, 1,
-            "Should remove exactly 1 file in first batch"
-        );
-        txn.remove_files(FilteredEngineData::try_new(data, selection_vector)?);
-
-        // Second batch: Remove only the last file
-        let scan3 = snapshot.clone().scan_builder().build()?;
-        let scan_metadata3 = scan3.scan_metadata(engine.as_ref())?.next().unwrap()?;
-        let (data2, mut selection_vector2) = scan_metadata3.scan_files.into_parts();
-
-        // Find the last selected file and keep only that one selected
-        let mut last_selected_idx = None;
-        for (i, &selected) in selection_vector2.iter().enumerate() {
-            if selected {
-                last_selected_idx = Some(i);
-            }
-        }
-
-        // Deselect all except the last one
-        for (i, selected) in selection_vector2.iter_mut().enumerate() {
-            if Some(i) != last_selected_idx {
-                *selected = false;
-            }
-        }
-
-        let second_batch_removed = selection_vector2.iter().filter(|&x| *x).count();
-        assert_eq!(
-            second_batch_removed, 1,
-            "Should remove exactly 1 file in second batch"
-        );
-        txn.remove_files(FilteredEngineData::try_new(data2, selection_vector2)?);
-
-        // Commit the transaction
-        let result = txn.commit(engine.as_ref())?;
-
-        match result {
-            CommitResult::CommittedTransaction(committed) => {
-                assert_eq!(committed.commit_version(), 6);
-
-                // Verify that exactly 2 files were removed (1 from each batch)
-                let new_snapshot = Snapshot::builder_for(table_url.clone())
-                    .at_version(6)
-                    .build(engine.as_ref())?;
-
-                let new_scan = new_snapshot.scan_builder().build()?;
-                let mut new_file_count = 0;
-                for new_metadata in new_scan.scan_metadata(engine.as_ref())? {
-                    let metadata = new_metadata?;
-                    new_file_count += metadata
-                        .scan_files
-                        .selection_vector()
-                        .iter()
-                        .filter(|&x| *x)
-                        .count();
-                }
-
-                // Verify we removed exactly 2 files (1 + 1)
-                let total_removed = first_batch_removed + second_batch_removed;
-                assert_eq!(total_removed, 2);
-                assert_eq!(new_file_count, initial_file_count - total_removed);
-                assert!(new_file_count > 0, "At least one file should remain");
-            }
-            _ => panic!("Transaction did not succeed"),
-        }
-    }
-    Ok(())
-}
-
-/// Regression test for https://github.com/delta-io/delta-kernel-rs/issues/2040
-///
-/// When `scan_metadata()` is called with a predicate, the scan row schema includes a
-/// `stats_parsed` column (7th column). Passing that scan metadata to `remove_files()` then
-/// `commit()` previously failed with "Too few fields in output schema" because the transform
-/// evaluator exhausted the output schema when it encountered the extra column.
-///
-/// Both predicate and non-predicate scans are tested because `remove_files` should behave
-/// identically regardless. Cases also vary the checkpoint format:
-/// - `use_struct_stats_checkpoint=false`: `stats` is non-null (raw JSON from the Add action). The
-///   remove action's `stats` comes from passthrough.
-/// - `use_struct_stats_checkpoint=true`: a checkpoint is written with `writeStatsAsJson=false,
-///   writeStatsAsStruct=true`, so the checkpoint stores `stats_parsed` but omits the raw `stats`
-///   JSON string. Scan rows from that checkpoint have `stats=null` but `stats_parsed` non-null. The
-///   remove action's `stats` is produced via `coalesce(null, to_json(stats_parsed))`, which
-///   exercises the coalesce path in the fix.
-#[rstest::rstest]
-#[case(false, false)]
-#[case(false, true)]
-#[case(true, false)]
-#[case(true, true)]
-// Multi-thread runtime required because the struct-stats checkpoint case calls
-// `Snapshot::checkpoint`, which makes nested `block_on` calls that deadlock
-// a single-threaded executor. `TokioMultiThreadExecutor` uses `block_in_place`
-// which avoids the deadlock.
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn test_remove_files_after_predicate_scan_includes_stats_parsed(
-    #[case] use_struct_stats_checkpoint: bool,
-    #[case] use_predicate: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    // Use a local directory so we can inspect the commit log for stats content.
-    let tmp_dir = tempdir()?;
-    let tmp_url = Url::from_directory_path(tmp_dir.path()).unwrap();
-
-    for (table_url, engine, _store, _table_name) in
-        setup_test_tables(schema.clone(), &[], Some(&tmp_url), "test_table").await?
-    {
-        let engine = Arc::new(engine);
-
-        // Write data (two parquet files with numbers [1,2,3] and [4,5,6]).
-        write_data_and_check_result_and_stats(table_url.clone(), schema.clone(), engine.clone(), 1)
-            .await?;
-
-        // When use_struct_stats_checkpoint=true, update table properties so the checkpoint
-        // omits the stats JSON string (writeStatsAsJson=false) but stores stats as a struct
-        // (writeStatsAsStruct=true). After checkpointing, scan rows from the checkpoint have
-        // stats=null and stats_parsed=non-null, exercising the coalesce path in the fix.
-        let snapshot = if use_struct_stats_checkpoint {
-            let table_path = table_url.to_file_path().unwrap();
-            let snapshot_v2 = set_table_properties(
-                table_path.to_str().unwrap(),
-                &table_url,
-                engine.as_ref(),
-                1,
-                &[
-                    ("delta.checkpoint.writeStatsAsJson", "false"),
-                    ("delta.checkpoint.writeStatsAsStruct", "true"),
-                ],
-            )?;
-            // `Snapshot::checkpoint` makes nested `block_on` calls internally (it reads the
-            // log segment lazily while writing). This requires `TokioMultiThreadExecutor`,
-            // which uses `block_in_place` to avoid deadlocking a single-thread runtime.
-            let mt_engine = create_default_engine_mt_executor(&table_url)?;
-            snapshot_v2.checkpoint(mt_engine.as_ref())?;
-            Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?
-        } else {
-            Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?
-        };
-
-        // commit_version = 2 (no checkpoint) or 3 (properties commit + checkpoint bump)
-        let expected_commit_version = if use_struct_stats_checkpoint { 3 } else { 2 };
-
-        // Always request all stats columns so stats_parsed is present in scan metadata
-        // regardless of whether a predicate is used. This ensures remove_files can always
-        // reconstruct stats (including the coalesce path when writeStatsAsJson=false).
-        let mut scan_builder = snapshot.clone().scan_builder().include_all_stats_columns();
-        if use_predicate {
-            scan_builder = scan_builder.with_predicate(Arc::new(Pred::gt(
-                column_expr!("number"),
-                Expr::literal(0_i32),
-            )));
-        }
-        let scan = scan_builder.build()?;
-
-        let mut txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-            .with_data_change(true);
-
-        // Pass scan metadata (which contains stats_parsed) directly to remove_files.
-        // This previously failed with "Too few fields in output schema".
-        for scan_metadata in scan.scan_metadata(engine.as_ref())? {
-            txn.remove_files(scan_metadata?.scan_files);
-        }
-
-        let committed = txn.commit(engine.as_ref())?.unwrap_committed();
-        assert_eq!(committed.commit_version(), expected_commit_version);
-
-        let remove_actions =
-            read_actions_from_commit(&table_url, expected_commit_version, "remove")?;
-        assert!(
-            !remove_actions.is_empty(),
-            "expected remove actions in commit"
-        );
-
-        // stats must be populated in every remove action: stats_parsed is always present
-        // (via include_all_stats_columns), so the coalesce path handles even checkpoints
-        // that omit the raw JSON stats string (writeStatsAsJson=false).
-        for remove in &remove_actions {
-            let stats_str = remove["stats"]
-                .as_str()
-                .expect("stats field should be a non-null JSON string");
-            let stats: serde_json::Value = serde_json::from_str(stats_str)?;
-            assert!(
-                stats["numRecords"].as_i64().unwrap_or(0) > 0,
-                "stats.numRecords should be populated, got: {stats}"
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Remove files via scan metadata on a partitioned table. Covers three predicate
-/// shapes against the same table so the remove-transform correctly handles the
-/// parsed scan columns in every combination:
-/// - no predicate: no `partitionValues_parsed`.
-/// - data-column predicate: no `partitionValues_parsed` (negative case; the fix must not affect
-///   scans whose predicate misses the partition columns).
-/// - partition predicate: `partitionValues_parsed` present.
-///
-/// Every case calls `.include_all_stats_columns()`, which forces `stats_parsed`
-/// into the scan output regardless of the predicate shape, so the partition-
-/// predicate case exercises both parsed-column drop paths together while the
-/// other two exercise only the `stats_parsed` drop path. The coalesce
-/// *reconstruction* of `stats` from `stats_parsed` is not exercised here
-/// because `stats` is non-null; the sibling
-/// `test_remove_files_after_predicate_scan_includes_stats_parsed` covers that.
-///
-/// `expected_partitions` is the multiset of `country` values expected across
-/// the generated Remove actions. Its length gives the expected Remove count,
-/// and its contents pin the correct partition was chosen (catches regressions
-/// where the wrong partition is removed).
-#[rstest::rstest]
-#[case::no_predicate(None, &["usa", "japan"])]
-#[case::data_predicate(
-    Some(Pred::gt(column_expr!("id"), Expr::literal(0_i32))),
-    &["usa", "japan"]
-)]
-#[case::partition_predicate(
-    Some(Pred::eq(column_expr!("country"), Expr::literal("usa".to_string()))),
-    &["usa"]
-)]
-#[tokio::test]
-async fn test_remove_files_partitioned_with_parsed_columns(
-    #[case] predicate: Option<Pred>,
-    #[case] expected_partitions: &[&str],
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let partition_col = "country";
-    let table_schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("id", DataType::INTEGER),
-        StructField::nullable("country", DataType::STRING),
-    ])?);
-    let data_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "id",
-        DataType::INTEGER,
-    )])?);
-
-    // Local directory backing: `read_actions_from_commit` reads commit JSON off disk
-    // and does not support the default in-memory store's `memory://` URL.
-    let tmp_dir = tempdir()?;
-    let tmp_url = Url::from_directory_path(tmp_dir.path()).unwrap();
-
-    for (table_url, engine, _store, _table_name) in setup_test_tables(
-        table_schema.clone(),
-        &[partition_col],
-        Some(&tmp_url),
-        "test_table",
-    )
-    .await?
-    {
-        let engine = Arc::new(engine);
-
-        // Write two partitions: country="usa" and country="japan".
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-        let mut txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-            .with_data_change(true);
-        let append_data = [[1, 2, 3], [10, 20, 30]].map(|data| -> DeltaResult<_> {
-            let data = RecordBatch::try_new(
-                Arc::new(data_schema.as_ref().try_into_arrow()?),
-                vec![Arc::new(Int32Array::from(data.to_vec()))],
-            )?;
-            Ok(Box::new(ArrowEngineData::new(data)))
-        });
-        for (data, partition_val) in append_data.into_iter().zip(["usa", "japan"]) {
-            let ctx = Arc::new(txn.partitioned_write_context(HashMap::from([(
-                partition_col.to_string(),
-                Scalar::String(partition_val.into()),
-            )]))?);
-            let add_meta = engine.write_parquet(data?.as_ref(), ctx.as_ref()).await?;
-            txn.add_files(add_meta);
-        }
-        txn.commit(engine.as_ref())?.unwrap_committed();
-
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-        let mut scan_builder = snapshot.clone().scan_builder().include_all_stats_columns();
-        if let Some(pred) = predicate.clone() {
-            scan_builder = scan_builder.with_predicate(Arc::new(pred));
-        }
-        let scan = scan_builder.build()?;
-
-        let mut txn = snapshot
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-            .with_data_change(true);
-        for scan_metadata in scan.scan_metadata(engine.as_ref())? {
-            txn.remove_files(scan_metadata?.scan_files);
-        }
-        let committed = txn.commit(engine.as_ref())?.unwrap_committed();
-        assert_eq!(committed.commit_version(), 2);
-
-        let remove_actions = read_actions_from_commit(&table_url, 2, "remove")?;
-        assert_eq!(
-            remove_actions.len(),
-            expected_partitions.len(),
-            "unexpected remove count; got {}: {remove_actions:?}",
-            remove_actions.len()
-        );
-
-        let mut actual_partitions: Vec<String> = remove_actions
-            .iter()
-            .filter_map(|r| {
-                r["partitionValues"][partition_col]
-                    .as_str()
-                    .map(String::from)
-            })
-            .collect();
-        actual_partitions.sort();
-        let mut expected_sorted: Vec<String> =
-            expected_partitions.iter().map(|s| s.to_string()).collect();
-        expected_sorted.sort();
-        assert_eq!(
-            actual_partitions, expected_sorted,
-            "partitionValues mismatch across removes; got: {remove_actions:?}"
-        );
-
-        // stats_parsed is present on every scan row, so the stats-with-parsed
-        // evaluator is selected for every case; it must still yield a populated
-        // stats JSON on every remove action.
-        for remove in &remove_actions {
-            let stats_str = remove["stats"]
-                .as_str()
-                .expect("stats field should be a non-null JSON string");
-            let stats: serde_json::Value = serde_json::from_str(stats_str)?;
-            assert!(
-                stats["numRecords"].as_i64().unwrap_or(0) > 0,
-                "stats.numRecords should be populated, got: {stats}"
-            );
-        }
-    }
-    Ok(())
-}
-
-// Helper function to create a table with CDF enabled
-async fn create_cdf_table(
-    table_name: &str,
-    schema: SchemaRef,
-) -> Result<(Url, Arc<DefaultEngine<TokioBackgroundExecutor>>, TempDir), Box<dyn std::error::Error>>
-{
-    let tmp_dir = tempdir()?;
-    let tmp_test_dir_url = Url::from_directory_path(tmp_dir.path()).unwrap();
-
-    let (store, engine, table_location) = engine_store_setup(table_name, Some(&tmp_test_dir_url));
-
-    let table_url = create_table(
-        store.clone(),
-        table_location,
-        schema.clone(),
-        &[],
-        true, // use protocol 3.7
-        vec![],
-        vec!["changeDataFeed"],
-    )
-    .await?;
-
-    Ok((table_url, Arc::new(engine), tmp_dir))
-}
-
-// Helper function to write data to a table
-async fn write_data_to_table(
-    table_url: &Url,
-    engine: &Arc<DefaultEngine<TokioBackgroundExecutor>>,
-    schema: SchemaRef,
-    values: Vec<i32>,
-) -> Result<Version, Box<dyn std::error::Error>> {
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test");
-
-    add_files_to_transaction(&mut txn, engine, schema, values).await?;
-
-    let result = txn.commit(engine.as_ref())?;
-    match result {
-        CommitResult::CommittedTransaction(committed) => Ok(committed.commit_version()),
-        _ => panic!("Transaction should be committed"),
-    }
-}
-
-// Helper function to add files to an existing transaction
-async fn add_files_to_transaction(
-    txn: &mut delta_kernel::transaction::Transaction,
-    engine: &Arc<DefaultEngine<TokioBackgroundExecutor>>,
-    schema: SchemaRef,
-    values: Vec<i32>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let data = RecordBatch::try_new(
-        Arc::new(schema.as_ref().try_into_arrow()?),
-        vec![Arc::new(Int32Array::from(values))],
-    )?;
-
-    let write_context = Arc::new(txn.unpartitioned_write_context().unwrap());
-    let add_files_metadata = engine
-        .write_parquet(&ArrowEngineData::new(data), write_context.as_ref())
-        .await?;
-    txn.add_files(add_files_metadata);
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_cdf_write_all_adds_succeeds() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that add-only transactions work with CDF enabled
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let (table_url, engine, _tmp_dir) =
-        create_cdf_table("test_cdf_all_adds", schema.clone()).await?;
-
-    // Add files - this should succeed
-    let version = write_data_to_table(&table_url, &engine, schema, vec![1, 2, 3]).await?;
-    assert_eq!(version, 1);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_cdf_write_all_removes_succeeds() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that remove-only transactions work with CDF enabled
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let (table_url, engine, _tmp_dir) =
-        create_cdf_table("test_cdf_all_removes", schema.clone()).await?;
-
-    // First, add some data
-    write_data_to_table(&table_url, &engine, schema, vec![1, 2, 3]).await?;
-
-    // Now remove the files
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("cdf remove test")
-        .with_data_change(true);
-
-    let scan = snapshot.scan_builder().build()?;
-    let scan_metadata = scan.scan_metadata(engine.as_ref())?.next().unwrap()?;
-    let (data, selection_vector) = scan_metadata.scan_files.into_parts();
-    txn.remove_files(FilteredEngineData::try_new(data, selection_vector)?);
-
-    // This should succeed - remove-only transactions are allowed with CDF
-    let result = txn.commit(engine.as_ref())?;
-    match result {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(committed.commit_version(), 2);
-        }
-        _ => panic!("Transaction should be committed"),
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_cdf_write_mixed_no_data_change_succeeds() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that mixed add+remove transactions work when dataChange=false.
-    // It's allowed because the transaction does not contain any logical data changes.
-    // This can happen when a table is being optimized/compacted.
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let (table_url, engine, _tmp_dir) =
-        create_cdf_table("test_cdf_mixed_no_data_change", schema.clone()).await?;
-
-    // First, add some data
-    write_data_to_table(&table_url, &engine, schema.clone(), vec![1, 2, 3]).await?;
-
-    // Now create a transaction with both add AND remove files, but dataChange=false
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("cdf mixed test")
-        .with_data_change(false); // dataChange=false is key here
-
-    // Add new files
-    add_files_to_transaction(&mut txn, &engine, schema, vec![4, 5, 6]).await?;
-
-    // Also remove existing files
-    let scan = snapshot.scan_builder().build()?;
-    let scan_metadata = scan.scan_metadata(engine.as_ref())?.next().unwrap()?;
-    let (data, selection_vector) = scan_metadata.scan_files.into_parts();
-    txn.remove_files(FilteredEngineData::try_new(data, selection_vector)?);
-
-    // This should succeed - mixed operations are allowed when dataChange=false
-    let result = txn.commit(engine.as_ref())?;
-    match result {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(committed.commit_version(), 2);
-        }
-        _ => panic!("Transaction should be committed"),
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_cdf_write_mixed_with_data_change_fails() -> Result<(), Box<dyn std::error::Error>> {
-    // This test verifies that mixed add+remove transactions fail with helpful error when
-    // dataChange=true
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = get_simple_int_schema();
-
-    let (table_url, engine, _tmp_dir) =
-        create_cdf_table("test_cdf_mixed_with_data_change", schema.clone()).await?;
-
-    // First, add some data
-    write_data_to_table(&table_url, &engine, schema.clone(), vec![1, 2, 3]).await?;
-
-    // Now create a transaction with both add AND remove files with dataChange=true
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("cdf mixed fail test")
-        .with_data_change(true); // dataChange=true - this should fail
-
-    // Add new files
-    add_files_to_transaction(&mut txn, &engine, schema, vec![4, 5, 6]).await?;
-
-    // Also remove existing files
-    let scan = snapshot.scan_builder().build()?;
-    let scan_metadata = scan.scan_metadata(engine.as_ref())?.next().unwrap()?;
-    let (data, selection_vector) = scan_metadata.scan_files.into_parts();
-    txn.remove_files(FilteredEngineData::try_new(data, selection_vector)?);
-
-    // This should fail with our new error message
-    assert_result_error_with_message(
-        txn.commit(engine.as_ref()),
-        "Cannot add and remove data in the same transaction when Change Data Feed is enabled (delta.enableChangeDataFeed = true). \
-         This would require writing CDC files for DML operations, which is not yet supported. \
-         Consider using separate transactions: one to add files, another to remove files."
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_post_commit_snapshot_create_then_insert() -> DeltaResult<()> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let temp_dir = tempdir().unwrap();
-    let table_url = Url::from_directory_path(temp_dir.path()).unwrap();
-    let engine = create_default_engine(&table_url)?;
-    let schema = get_simple_int_schema();
-
-    // Create table and verify post_commit_snapshot
-    let create_result = create_table_txn(table_url.as_str(), schema, env!("CARGO_PKG_VERSION"))
-        .build(engine.as_ref(), Box::new(FileSystemCommitter::new()))?
-        .commit(engine.as_ref())?;
-
-    let mut current_snapshot = match create_result {
-        CommitResult::CommittedTransaction(committed) => {
-            assert_eq!(committed.commit_version(), 0);
-            // CREATE TABLE is the first commit: 1 commit since last checkpoint/compaction
-            assert_eq!(committed.post_commit_stats().commits_since_checkpoint, 1);
-            assert_eq!(
-                committed.post_commit_stats().commits_since_log_compaction,
-                1
-            );
-            let post_snapshot = committed
-                .post_commit_snapshot()
-                .expect("should have post_commit_snapshot");
-            assert_eq!(post_snapshot.version(), 0);
-            post_snapshot.clone()
-        }
-        _ => panic!("Create should succeed"),
-    };
-
-    // Do 10 inserts and verify post_commit_snapshot for each
-    for i in 1..11 {
-        let base_version = current_snapshot.version();
-
-        let txn = current_snapshot
-            .clone()
-            .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-            .with_engine_info("test");
-
-        match txn.commit(engine.as_ref())? {
-            CommitResult::CommittedTransaction(committed) => {
-                let post_snapshot = committed
-                    .post_commit_snapshot()
-                    .expect("should have post_commit_snapshot");
-
-                assert_eq!(post_snapshot.version(), base_version + 1);
-                assert_eq!(post_snapshot.version(), committed.commit_version());
-                assert_eq!(post_snapshot.schema(), current_snapshot.schema());
-                assert_eq!(post_snapshot.table_root(), current_snapshot.table_root());
-
-                current_snapshot = post_snapshot.clone();
-            }
-            _ => panic!("Commit {i} should succeed"),
-        }
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_write_parquet_succeed_with_logical_partition_names(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("id", DataType::INTEGER),
-        StructField::nullable("letter", DataType::STRING),
-    ])?);
-
-    for (table_url, engine, _store, _table_name) in setup_test_tables(
-        schema.clone(),
-        &["letter"],
-        None,
-        "test_partition_translate",
-    )
-    .await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-
-        // Create data with only the non-partition column
-        let data_schema = Arc::new(
-            StructType::try_new(vec![StructField::nullable("id", DataType::INTEGER)]).unwrap(),
-        );
-        let batch = RecordBatch::try_new(
-            Arc::new(data_schema.as_ref().try_into_arrow()?),
-            vec![Arc::new(Int32Array::from(vec![1, 2]))],
-        )?;
-
-        // Pass partition values with logical name — should succeed
-        let result = write_batch_to_table(
-            &snapshot,
-            &engine,
-            batch,
-            HashMap::from([("letter".to_string(), Scalar::String("a".into()))]),
-        )
-        .await;
-        assert!(
-            result.is_ok(),
-            "write_parquet should succeed with valid logical partition name"
-        );
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_write_parquet_rejects_partitioned_write_context_on_unpartitioned_table(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let schema = get_simple_int_schema();
-
-    for (table_url, engine, _store, _table_name) in
-        setup_test_tables(schema.clone(), &[], None, "test_partition_reject").await?
-    {
-        let snapshot = Snapshot::builder_for(table_url.clone()).build(&engine)?;
-        let txn = snapshot
-            .clone()
-            .transaction(Box::new(FileSystemCommitter::new()), &engine)?
-            .with_engine_info("test");
-
-        let result = txn.partitioned_write_context(HashMap::from([(
-            "nonexistent".to_string(),
-            Scalar::String("val".into()),
-        )]));
-        let err =
-            result.expect_err("should fail with partitioned_write_context on unpartitioned table");
-        let err_msg = err.to_string();
-        assert!(
-            err_msg.contains("table is not partitioned"),
-            "Error should indicate table is not partitioned, got: {err_msg}"
-        );
-    }
-    Ok(())
-}
-
-/// 1. Creates a table with the given column mapping mode
-/// 2. Writes two batches of data
-/// 3. Checkpoints and verifies add.stats uses physical column names in the checkpoint
-/// 4. Reads a parquet footer to verify physical names/IDs
-/// 5. Reads data back to verify correctness
-/// 6. Removes files and verifies remove.stats matches the original add.stats
-#[rstest::rstest]
-#[case::cm_none(ColumnMappingMode::None)]
-#[case::cm_id(ColumnMappingMode::Id)]
-#[case::cm_name(ColumnMappingMode::Name)]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_column_mapping_write(
-    #[case] cm_mode: ColumnMappingMode,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = nested_schema()?;
-
-    let (_tmp_dir, table_path, _) = test_table_setup()?;
-    let table_url = Url::from_directory_path(&table_path).unwrap();
-    let store: Arc<DynObjectStore> = Arc::new(LocalFileSystem::new());
-    let engine = Arc::new(
-        DefaultEngineBuilder::new(store.clone())
-            .with_task_executor(Arc::new(TokioMultiThreadExecutor::new(
-                tokio::runtime::Handle::current(),
-            )))
-            .build(),
-    );
-
-    // Step 1: Create table
-    let mode_str = match cm_mode {
-        ColumnMappingMode::None => "none",
-        ColumnMappingMode::Id => "id",
-        ColumnMappingMode::Name => "name",
-    };
-    let mut latest_snapshot = create_table_and_load_snapshot(
-        &table_path,
-        schema.clone(),
-        engine.as_ref(),
-        &[("delta.columnMapping.mode", mode_str)],
-    )?;
-
-    // Get physical field paths for stats verification (top-level and nested)
-    let cm = latest_snapshot
-        .table_properties()
-        .column_mapping_mode
-        .unwrap_or(ColumnMappingMode::None);
-    let row_number_physical = get_any_level_column_physical_name(
-        latest_snapshot.schema().as_ref(),
-        &ColumnName::new(["row_number"]),
-        cm,
-    )?
-    .into_inner();
-    let street_physical = get_any_level_column_physical_name(
-        latest_snapshot.schema().as_ref(),
-        &ColumnName::new(["address", "street"]),
-        cm,
-    )?
-    .into_inner();
-
-    // Step 2: Write two batches
-    for data in nested_batches()? {
-        latest_snapshot =
-            write_batch_to_table(&latest_snapshot, engine.as_ref(), data, HashMap::new()).await?;
-    }
-
-    // Enable writeStatsAsStruct so the checkpoint contains native stats_parsed.
-    // CREATE TABLE doesn't allow this property yet, so we write a metadata-update commit directly.
-    latest_snapshot = set_table_properties(
-        &table_path,
-        &table_url,
-        engine.as_ref(),
-        latest_snapshot.version(),
-        &[("delta.checkpoint.writeStatsAsStruct", "true")],
-    )?;
-
-    // Step 3: Checkpoint and verify add.stats uses correct column names
-    let snapshot_for_checkpoint = latest_snapshot.clone();
-    snapshot_for_checkpoint.checkpoint(engine.as_ref())?;
-    let ckpt_snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let add_actions = read_add_infos(&ckpt_snapshot, engine.as_ref())?;
-    let mut all_stats: Vec<_> = add_actions
-        .iter()
-        .filter_map(|a| a.stats.as_ref())
-        .filter(|s| s.get("minValues").is_some())
-        .collect();
-    assert_eq!(all_stats.len(), 2, "should have stats for 2 files");
-    all_stats.sort_by_key(|s| s["minValues"][&row_number_physical[0]].as_i64().unwrap());
-
-    // Batch 1: row_number 1..3, address.street "st1".."st3"
-    assert_min_max_stats(all_stats[0], &row_number_physical, 1, 3);
-    assert_min_max_stats(all_stats[0], &street_physical, "st1", "st3");
-
-    // Batch 2: row_number 4..6, address.street "st4".."st6"
-    assert_min_max_stats(all_stats[1], &row_number_physical, 4, 6);
-    assert_min_max_stats(all_stats[1], &street_physical, "st4", "st6");
-
-    // Step 3b: Verify stats_parsed in scan metadata uses correct physical column names
-    {
-        let scan = ckpt_snapshot
-            .scan_builder()
-            .include_all_stats_columns()
-            .build()?;
-        let scan_metadata_results: Vec<_> = scan
-            .scan_metadata(engine.as_ref())?
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let mut stats_rows: Vec<(i64, i64, String, String)> = Vec::new();
-        for sm in scan_metadata_results {
-            let (data, sel) = sm.scan_files.into_parts();
-            let batch: RecordBatch = ArrowEngineData::try_from_engine_data(data)?.into();
-
-            let batch_struct = StructArray::from(batch.clone());
-            let stats_parsed: &StructArray =
-                resolve_struct_field(&batch_struct, &["stats_parsed".into()]);
-
-            let min_path = |field: &[String]| -> Vec<String> {
-                [&["stats_parsed".into(), "minValues".into()], field].concat()
-            };
-            let max_path = |field: &[String]| -> Vec<String> {
-                [&["stats_parsed".into(), "maxValues".into()], field].concat()
-            };
-            let min_row_num: &Int64Array =
-                resolve_struct_field(&batch_struct, &min_path(&row_number_physical));
-            let max_row_num: &Int64Array =
-                resolve_struct_field(&batch_struct, &max_path(&row_number_physical));
-            let min_st: &StringArray =
-                resolve_struct_field(&batch_struct, &min_path(&street_physical));
-            let max_st: &StringArray =
-                resolve_struct_field(&batch_struct, &max_path(&street_physical));
-
-            for (i, &selected) in sel.iter().enumerate().take(batch.num_rows()) {
-                if selected && !stats_parsed.is_null(i) {
-                    stats_rows.push((
-                        min_row_num.value(i),
-                        max_row_num.value(i),
-                        min_st.value(i).to_string(),
-                        max_st.value(i).to_string(),
-                    ));
-                }
-            }
-        }
-
-        stats_rows.sort_by_key(|r| r.0);
-        assert_eq!(stats_rows.len(), 2, "should have stats_parsed for 2 files");
-        assert_eq!(stats_rows[0], (1, 3, "st1".to_string(), "st3".to_string()));
-        assert_eq!(stats_rows[1], (4, 6, "st4".to_string(), "st6".to_string()));
-    }
-
-    // Step 4: Read parquet footer to verify physical names and native field_id
-    {
-        let parquet_path = &add_actions
-            .first()
-            .expect("should have at least one add file")
-            .path;
-        let parquet_url = table_url.join(parquet_path)?;
-        let local_path = parquet_url.to_file_path().unwrap();
-
-        let obj_meta = store
-            .head(&Path::from_url_path(parquet_url.path())?)
-            .await?;
-        let file_meta = FileMeta::new(
-            parquet_url,
-            0, /* last_modified */
-            obj_meta.size as u64,
-        );
-        let footer = engine.parquet_handler().read_parquet_footer(&file_meta)?;
-        let footer_schema = footer.schema;
-
-        let logical_schema = latest_snapshot.schema();
-        for logical_path in [&["row_number"][..], &["address", "street"]] {
-            let col = ColumnName::new(logical_path.iter().copied());
-            let physical =
-                get_any_level_column_physical_name(logical_schema.as_ref(), &col, cm)?.into_inner();
-            assert_schema_has_field(&footer_schema, &physical);
-
-            let field_id = get_parquet_field_id(&local_path, &physical);
-            let logical_field = resolve_field(logical_schema.as_ref(), logical_path).unwrap();
-            match cm_mode {
-                ColumnMappingMode::Id | ColumnMappingMode::Name => {
-                    let expected_id =
-                        match logical_field.get_config_value(&ColumnMetadataKey::ColumnMappingId) {
-                            Some(MetadataValue::Number(n)) => *n as i32,
-                            other => panic!("expected ColumnMappingId number, got {other:?}"),
-                        };
-                    assert_eq!(
-                        field_id,
-                        Some(expected_id),
-                        "parquet field_id mismatch for {logical_path:?}"
-                    );
-                }
-                ColumnMappingMode::None => {
-                    assert_eq!(
-                        field_id, None,
-                        "parquet field_id should not be set in None column mapping mode"
-                    );
-                }
-            }
-        }
-    }
-
-    // Step 5: Read data back to verify correctness
-    {
-        let post_ckpt_snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-        let scan = post_ckpt_snapshot.scan_builder().build()?;
-        let batches: Vec<RecordBatch> = scan
-            .execute(engine.clone())?
-            .map(|r| {
-                let data = r.unwrap();
-                let arrow = ArrowEngineData::try_from_engine_data(data).unwrap();
-                arrow.record_batch().clone()
-            })
-            .collect();
-
-        let result_schema = batches[0].schema();
-        let combined = delta_kernel::arrow::compute::concat_batches(&result_schema, &batches)?;
-        assert_eq!(
-            combined.num_rows(),
-            6,
-            "Should have 6 rows from two written batches"
-        );
-
-        // Verify logical column names and data values
-        let combined_struct = StructArray::from(combined);
-
-        // Top-level: row_number should contain [1..=6]
-        let row_numbers: &Int64Array =
-            resolve_struct_field(&combined_struct, &["row_number".into()]);
-        let mut vals: Vec<i64> = (0..row_numbers.len())
-            .map(|i| row_numbers.value(i))
-            .collect();
-        vals.sort();
-        assert_eq!(vals, vec![1, 2, 3, 4, 5, 6]);
-
-        // Nested: address.street should contain ["st1"..="st6"]
-        let streets: &StringArray =
-            resolve_struct_field(&combined_struct, &["address".into(), "street".into()]);
-        let mut street_vals: Vec<&str> = (0..streets.len()).map(|i| streets.value(i)).collect();
-        street_vals.sort();
-        assert_eq!(street_vals, vec!["st1", "st2", "st3", "st4", "st5", "st6"]);
-    }
-
-    // Step 6: Remove files and verify remove.stats matches original add.stats
-    {
-        let original_add_stats: Vec<serde_json::Value> =
-            add_actions.iter().filter_map(|a| a.stats.clone()).collect();
-        assert!(
-            !original_add_stats.is_empty(),
-            "should have at least one add with stats"
-        );
-
-        let remove_actions =
-            remove_all_and_get_remove_actions(&latest_snapshot, &table_url, engine.as_ref())?;
-        assert!(
-            !remove_actions.is_empty(),
-            "Expected at least one remove action"
-        );
-
-        let remove_stats: Vec<serde_json::Value> = remove_actions
-            .iter()
-            .filter_map(|r| {
-                r["stats"]
-                    .as_str()
-                    .map(|s| serde_json::from_str(s).unwrap())
-            })
-            .collect();
-        assert_eq!(
-            remove_stats, original_add_stats,
-            "remove.stats should match original add.stats"
-        );
-    }
-
-    Ok(())
-}
-
-/// Verifies that partitioned writes use physical column names in add.partitionValues.
-#[rstest::rstest]
-#[case::cm_none("./tests/data/partition_cm/none")]
-#[case::cm_id("./tests/data/partition_cm/id")]
-#[case::cm_name("./tests/data/partition_cm/name")]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_column_mapping_partitioned_write(
-    #[case] table_dir: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // Copy test data to a temp dir so we can write to it
-    let tmp_dir = tempdir()?;
-    copy_directory(std::path::Path::new(table_dir), tmp_dir.path())?;
-    let table_url = Url::from_directory_path(tmp_dir.path()).unwrap();
-    let store: Arc<DynObjectStore> = Arc::new(LocalFileSystem::new());
-    let engine = Arc::new(
-        DefaultEngineBuilder::new(store.clone())
-            .with_task_executor(Arc::new(TokioMultiThreadExecutor::new(
-                tokio::runtime::Handle::current(),
-            )))
-            .build(),
-    );
-
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let cm = snapshot
-        .table_properties()
-        .column_mapping_mode
-        .unwrap_or(ColumnMappingMode::None);
-    let physical_name = get_any_level_column_physical_name(
-        snapshot.schema().as_ref(),
-        &ColumnName::new(["category"]),
-        cm,
-    )?
-    .into_inner()
-    .remove(0);
-
-    // Verify physical name for column mapping mode
-    if table_dir.ends_with("none") {
-        assert_eq!(physical_name, "category");
-    } else {
-        assert_ne!(
-            physical_name, "category",
-            "physical name should differ from logical name under column mapping"
-        );
-    }
-
-    // Write data with partition value
-    let data_schema = Arc::new(StructType::try_new(vec![StructField::nullable(
-        "value",
-        DataType::INTEGER,
-    )])?);
-    let batch = RecordBatch::try_new(
-        Arc::new(data_schema.as_ref().try_into_arrow()?),
-        vec![Arc::new(Int32Array::from(vec![1, 2]))],
-    )?;
-    let partition_values = HashMap::from([("category".to_string(), Scalar::String("A".into()))]);
-    write_batch_to_table(&snapshot, engine.as_ref(), batch, partition_values).await?;
-
-    // Read commit log and verify add.partitionValues key uses physical name
-    let add_actions = read_actions_from_commit(&table_url, 1, "add")?;
-    assert!(!add_actions.is_empty(), "no add action found in commit log");
-    for add in &add_actions {
-        assert_partition_values(add, &physical_name, "A");
-    }
-
-    // Remove the written file and verify remove action preserves physical names
-    let post_write_snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let remove_actions =
-        remove_all_and_get_remove_actions(&post_write_snapshot, &table_url, engine.as_ref())?;
-    assert!(
-        !remove_actions.is_empty(),
-        "no remove action found in commit log"
-    );
-    for remove in &remove_actions {
-        assert_partition_values(remove, &physical_name, "A");
-    }
-
-    Ok(())
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn test_checkpoint_non_kernel_written_table() {
-    // Table written by a non-kernel-integrated connector: 7 rows with columns (i, j, k), where
-    // parquet field nullabilities differ from the Delta schema. DefaultEngine reads it, coerces
-    // nullabilities to match the Delta schema, creates a checkpoint, and verifies the data is
-    // unchanged.
-    let source_path = std::path::Path::new("./tests/data/external-table-different-nullability");
-    let temp_dir = tempfile::tempdir().unwrap();
-    let table_path = temp_dir.path().join("test-checkpoint-table");
-    test_utils::copy_directory(source_path, &table_path).unwrap();
-
-    let url = Url::from_directory_path(&table_path).unwrap();
-    let store: Arc<DynObjectStore> = Arc::new(LocalFileSystem::new());
-    let executor = Arc::new(
-        delta_kernel::engine::default::executor::tokio::TokioMultiThreadExecutor::new(
-            tokio::runtime::Handle::current(),
-        ),
-    );
-    let engine: Arc<delta_kernel::engine::default::DefaultEngine<_>> = Arc::new(
-        delta_kernel::engine::default::DefaultEngineBuilder::new(store)
-            .with_task_executor(executor)
-            .build(),
-    );
-
-    // Read data before checkpoint
-    let snapshot = Snapshot::builder_for(url.clone())
-        .build(engine.as_ref())
-        .unwrap();
-    let scan_before = Arc::clone(&snapshot).scan_builder().build().unwrap();
-    let batches_before = test_utils::read_scan(&scan_before, engine.clone()).unwrap();
-
-    // Create checkpoint via snapshot.checkpoint()
-    snapshot.checkpoint(engine.as_ref()).unwrap();
-
-    // Read data after checkpoint
-    let snapshot_after = Snapshot::builder_for(url.clone())
-        .build(engine.as_ref())
-        .unwrap();
-    let scan_after = snapshot_after.scan_builder().build().unwrap();
-    let batches_after = test_utils::read_scan(&scan_after, engine.clone()).unwrap();
-
-    // Verify data unchanged
-    let formatted_before =
-        delta_kernel::arrow::util::pretty::pretty_format_batches(&batches_before)
-            .unwrap()
-            .to_string();
-    let formatted_after = delta_kernel::arrow::util::pretty::pretty_format_batches(&batches_after)
-        .unwrap()
-        .to_string();
-    assert_eq!(
-        formatted_before, formatted_after,
-        "Row data changed after checkpoint creation!"
-    );
-
-    // Verify checkpoint file exists
-    let delta_log_path = table_path.join("_delta_log");
-    let has_checkpoint = std::fs::read_dir(&delta_log_path)
-        .unwrap()
-        .filter_map(|e| e.ok())
-        .any(|e| {
-            e.file_name()
-                .to_str()
-                .is_some_and(|n| n.contains(".checkpoint.parquet"))
-        });
-    assert!(has_checkpoint, "Expected at least one checkpoint file");
-}
-
-struct ClusteredTableSetup {
-    _tmp_dir: TempDir,
-    table_path: String,
-    table_url: Url,
-    engine: Arc<DefaultEngine<TokioMultiThreadExecutor>>,
-    snapshot: Arc<Snapshot>,
-}
-
-/// Creates a clustered table with column mapping and sets table properties.
-fn setup_clustered_table(
-    cm_mode: &str,
-    schema: Arc<StructType>,
-    clustering_cols: Vec<ColumnName>,
-    table_properties: &[(&str, &str)],
-) -> Result<ClusteredTableSetup, Box<dyn std::error::Error>> {
-    use delta_kernel::transaction::data_layout::DataLayout;
-
-    let (_tmp_dir, table_path, _) = test_table_setup()?;
-    let table_url = Url::from_directory_path(&table_path).unwrap();
-    let engine = create_default_engine_mt_executor(&table_url)?;
-
-    let _ = create_table_txn(table_url.as_str(), schema, "Test/1.0")
-        .with_table_properties([("delta.columnMapping.mode", cm_mode)])
-        .with_data_layout(DataLayout::Clustered {
-            columns: clustering_cols,
-        })
-        .build(engine.as_ref(), Box::new(FileSystemCommitter::new()))?
-        .commit(engine.as_ref())?;
-
-    let snapshot = set_table_properties(
-        &table_path,
-        &table_url,
-        engine.as_ref(),
-        0,
-        table_properties,
-    )?;
-
-    Ok(ClusteredTableSetup {
-        _tmp_dir,
-        table_path,
-        table_url,
-        engine,
-        snapshot,
-    })
-}
-
-/// E2E test: create a clustered table with column mapping, write data, and verify that
-/// add.stats in the commit log contains min/max statistics for the clustering columns
-/// (including a nested column).
-#[rstest::rstest]
-#[case::cm_none("none")]
-#[case::cm_name("name")]
-#[case::cm_id("id")]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_clustered_table_write_has_stats(
-    #[case] cm_mode: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let clustering_cols = vec![
-        ColumnName::new(["row_number"]),
-        ColumnName::new(["address", "street"]),
-    ];
-    let setup = setup_clustered_table(
-        cm_mode,
-        nested_schema()?,
-        clustering_cols.clone(),
-        &[("delta.dataSkippingNumIndexedCols", "0")],
-    )?;
-    let engine = &setup.engine;
-    let mut snapshot = setup.snapshot;
-    for batch in nested_batches()? {
-        snapshot = write_batch_to_table(&snapshot, engine.as_ref(), batch, HashMap::new()).await?;
-    }
-
-    let cm = assert_column_mapping_mode(&snapshot, cm_mode);
-    let physical_paths: Vec<Vec<String>> = clustering_cols
-        .iter()
-        .map(|c| {
-            get_any_level_column_physical_name(snapshot.schema().as_ref(), c, cm)
-                .unwrap()
-                .into_inner()
-        })
-        .collect();
-    if cm != ColumnMappingMode::None {
-        let logical_paths: Vec<Vec<&str>> = vec![vec!["row_number"], vec!["address", "street"]];
-        for (phys, logical) in physical_paths.iter().zip(&logical_paths) {
-            assert_ne!(
-                phys.iter().map(String::as_str).collect_vec(),
-                *logical,
-                "physical path should differ from logical when cm={cm:?}"
-            );
-        }
-    }
-
-    // Resolve a non-clustering column to verify it's excluded from stats
-    let non_clustering_physical = get_any_level_column_physical_name(
-        snapshot.schema().as_ref(),
-        &ColumnName::new(["name"]),
-        cm,
-    )?
-    .into_inner();
-
-    // Verify stats for each write commit (v2 and v3, since v1 is the property update).
-    // Batch 1 (v2): row_number 1..3, address.street "st1".."st3"
-    // Batch 2 (v3): row_number 4..6, address.street "st4".."st6"
-    let expected: [(i64, i64, &str, &str); 2] = [(1, 3, "st1", "st3"), (4, 6, "st4", "st6")];
-    for (version, (min_rn, max_rn, min_st, max_st)) in expected.iter().enumerate() {
-        let version = (version + 2) as u64;
-        let add_actions = read_actions_from_commit(&setup.table_url, version, "add")?;
-        assert!(
-            !add_actions.is_empty(),
-            "v{version}: should have add actions"
-        );
-
-        for add in &add_actions {
-            let stats: serde_json::Value = serde_json::from_str(
-                add.get("stats")
-                    .and_then(|s| s.as_str())
-                    .expect("add action should have stats"),
-            )?;
-            // Clustering columns should have stats despite numIndexedCols=0
-            assert_min_max_stats(&stats, &physical_paths[0], *min_rn, *max_rn);
-            assert_min_max_stats(&stats, &physical_paths[1], *min_st, *max_st);
-
-            // Non-clustering column "name" should NOT have stats
-            let non_cluster_min = resolve_json_path(&stats["minValues"], &non_clustering_physical);
-            assert!(
-                non_cluster_min.is_null(),
-                "v{version}: non-clustering column 'name' should not have stats"
-            );
-        }
-    }
-
-    Ok(())
-}
-
-/// E2E test: create a clustered table with column mapping, enable writeStatsAsStruct,
-/// write data, checkpoint, and verify stats_parsed.
-#[rstest::rstest]
-#[case::cm_none("none")]
-#[case::cm_name("name")]
-#[case::cm_id("id")]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_clustered_table_write_has_stats_parsed(
-    #[case] cm_mode: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let clustering_cols = vec![
-        ColumnName::new(["row_number"]),
-        ColumnName::new(["address", "street"]),
-    ];
-    let setup = setup_clustered_table(
-        cm_mode,
-        nested_schema()?,
-        clustering_cols.clone(),
-        &[
-            ("delta.checkpoint.writeStatsAsStruct", "true"),
-            ("delta.dataSkippingNumIndexedCols", "0"),
-        ],
-    )?;
-    let engine = &setup.engine;
-    let mut snapshot = setup.snapshot;
-    for batch in nested_batches()? {
-        snapshot = write_batch_to_table(&snapshot, engine.as_ref(), batch, HashMap::new()).await?;
-    }
-
-    let cm = assert_column_mapping_mode(&snapshot, cm_mode);
-    let physical_paths: Vec<Vec<String>> = clustering_cols
-        .iter()
-        .map(|c| {
-            get_any_level_column_physical_name(snapshot.schema().as_ref(), c, cm)
-                .unwrap()
-                .into_inner()
-        })
-        .collect();
-    if cm != ColumnMappingMode::None {
-        let logical_paths: Vec<Vec<&str>> = vec![vec!["row_number"], vec!["address", "street"]];
-        for (phys, logical) in physical_paths.iter().zip(&logical_paths) {
-            assert_ne!(
-                phys.iter().map(String::as_str).collect_vec(),
-                *logical,
-                "physical path should differ from logical when cm={cm:?}"
-            );
-        }
-    }
-    let non_clustering_physical = get_any_level_column_physical_name(
-        snapshot.schema().as_ref(),
-        &ColumnName::new(["name"]),
-        cm,
-    )?
-    .into_inner();
-
-    snapshot.checkpoint(engine.as_ref())?;
-
-    // Read checkpoint parquet directly to verify stats_parsed contains only clustering columns.
-    // ScanBuilder::include_all_stats_columns() doesn't support stats_parsed when
-    // dataSkippingNumIndexedCols=0. Read directly from the checkpoint parquet file instead.
-    use delta_kernel::parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-    let delta_log = std::path::Path::new(&setup.table_path).join("_delta_log");
-    let ckpt_path = std::fs::read_dir(&delta_log)?
-        .filter_map(|e| e.ok())
-        .find(|e| {
-            e.file_name()
-                .to_str()
-                .is_some_and(|n| n.contains(".checkpoint.parquet"))
-        })
-        .expect("checkpoint parquet should exist")
-        .path();
-    let file = std::fs::File::open(&ckpt_path)?;
-    let reader = ParquetRecordBatchReaderBuilder::try_new(file)?.build()?;
-
-    let min_path = |field: &[String]| -> Vec<String> { [&["minValues".into()], field].concat() };
-    let max_path = |field: &[String]| -> Vec<String> { [&["maxValues".into()], field].concat() };
-
-    let mut stats_rows: Vec<(i64, i64, String, String)> = Vec::new();
-    for batch in reader {
-        let batch = batch?;
-        let batch_struct = StructArray::from(batch);
-        let add: &StructArray = resolve_struct_field(&batch_struct, &["add".into()]);
-        let stats_parsed: &StructArray = resolve_struct_field(add, &["stats_parsed".into()]);
-
-        // Non-clustering column should not appear in stats_parsed
-        let min_values: &StructArray = resolve_struct_field(stats_parsed, &["minValues".into()]);
-        assert!(
-            min_values
-                .column_by_name(&non_clustering_physical[0])
-                .is_none(),
-            "non-clustering column '{}' should not have stats_parsed",
-            non_clustering_physical[0]
-        );
-
-        let min_row_num: &Int64Array =
-            resolve_struct_field(stats_parsed, &min_path(&physical_paths[0]));
-        let max_row_num: &Int64Array =
-            resolve_struct_field(stats_parsed, &max_path(&physical_paths[0]));
-        let min_st: &StringArray =
-            resolve_struct_field(stats_parsed, &min_path(&physical_paths[1]));
-        let max_st: &StringArray =
-            resolve_struct_field(stats_parsed, &max_path(&physical_paths[1]));
-
-        for i in 0..stats_parsed.len() {
-            if !stats_parsed.is_null(i) {
-                stats_rows.push((
-                    min_row_num.value(i),
-                    max_row_num.value(i),
-                    min_st.value(i).to_string(),
-                    max_st.value(i).to_string(),
-                ));
-            }
-        }
-    }
-
-    stats_rows.sort_by_key(|r| r.0);
-    assert_eq!(stats_rows.len(), 2, "should have stats_parsed for 2 files");
-    assert_eq!(stats_rows[0], (1, 3, "st1".to_string(), "st3".to_string()));
-    assert_eq!(stats_rows[1], (4, 6, "st4".to_string(), "st6".to_string()));
-
-    Ok(())
-}
-
-// === Path format tests ===
-
-fn get_simple_schema() -> SchemaRef {
-    Arc::new(StructType::try_new(vec![StructField::new("id", DataType::INTEGER, true)]).unwrap())
-}
-
-fn simple_id_batch(schema: &SchemaRef, values: Vec<i32>) -> RecordBatch {
-    RecordBatch::try_new(
-        Arc::new(schema.as_ref().try_into_arrow().unwrap()),
-        vec![Arc::new(Int32Array::from(values))],
-    )
-    .unwrap()
-}
-
-/// Helper to write a batch and return the post-commit snapshot.
-async fn write_batch_to_table_simple(
-    snapshot: &Arc<Snapshot>,
-    engine: &DefaultEngine<TokioBackgroundExecutor>,
-    data: RecordBatch,
-) -> Result<Arc<Snapshot>, Box<dyn std::error::Error>> {
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine)?
-        .with_engine_info("test");
-    let write_context = txn.unpartitioned_write_context()?;
-    let add_meta = engine
-        .write_parquet(&ArrowEngineData::new(data), &write_context)
-        .await?;
-    txn.add_files(add_meta);
-    let committed = txn.commit(engine)?.unwrap_committed();
-    Ok(committed.post_commit_snapshot().unwrap().clone())
-}
-
-#[tokio::test]
-async fn test_write_uses_relative_paths_and_readback() -> Result<(), Box<dyn std::error::Error>> {
-    let schema = get_simple_schema();
-    let (_tmp_dir, table_path, engine) = test_table_setup()?;
-    let snapshot =
-        create_table_and_load_snapshot(&table_path, schema.clone(), engine.as_ref(), &[])?;
-
-    let snapshot = write_batch_to_table_simple(
-        &snapshot,
-        engine.as_ref(),
-        simple_id_batch(&schema, vec![1, 2, 3]),
-    )
-    .await?;
-
-    // Verify paths in the log are relative (no scheme like "s3://")
-    let add_infos = read_add_infos(&snapshot, engine.as_ref())?;
-    assert_eq!(add_infos.len(), 1);
-    let path = &add_infos[0].path;
-    assert!(
-        !path.contains("://"),
-        "should produce relative paths, got: {path}"
-    );
-
-    // Verify data is readable via scan
-    let scan = snapshot.scan_builder().build()?;
-    let engine_ref: Arc<dyn Engine> = engine;
-    let batches = test_utils::read_scan(&scan, engine_ref)?;
-    assert_eq!(batches.len(), 1);
-    assert_eq!(batches[0].num_rows(), 3);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_multiple_files_in_commit_all_use_relative_paths(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let schema = get_simple_schema();
-    let (_tmp_dir, table_path, engine) = test_table_setup()?;
-    let snapshot =
-        create_table_and_load_snapshot(&table_path, schema.clone(), engine.as_ref(), &[])?;
-
-    let mut txn = snapshot
-        .clone()
-        .transaction(Box::new(FileSystemCommitter::new()), engine.as_ref())?
-        .with_engine_info("test");
-    let write_context = txn.unpartitioned_write_context().unwrap();
-    for values in [vec![1, 2], vec![3, 4]] {
-        let add_meta = engine
-            .write_parquet(
-                &ArrowEngineData::new(simple_id_batch(&schema, values)),
-                &write_context,
-            )
-            .await?;
-        txn.add_files(add_meta);
-    }
-    let committed = txn.commit(engine.as_ref())?.unwrap_committed();
-    let snapshot = committed.post_commit_snapshot().unwrap().clone();
-
-    let add_infos = read_add_infos(&snapshot, engine.as_ref())?;
-    assert_eq!(add_infos.len(), 2);
-    for info in &add_infos {
-        assert!(
-            !info.path.contains("://"),
-            "Expected relative path, got: {}",
-            info.path
-        );
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_multiple_commits_with_relative_paths_all_readable(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let schema = get_simple_schema();
-    let (_tmp_dir, table_path, engine) = test_table_setup()?;
-    let mut snapshot =
-        create_table_and_load_snapshot(&table_path, schema.clone(), engine.as_ref(), &[])?;
-
-    for values in [vec![1, 2], vec![3, 4], vec![5, 6]] {
-        snapshot = write_batch_to_table(
-            &snapshot,
-            engine.as_ref(),
-            simple_id_batch(&schema, values),
-            HashMap::new(),
-        )
-        .await?;
-    }
-
-    let scan = snapshot.scan_builder().build()?;
-    let engine_ref: Arc<dyn Engine> = engine;
-    let batches = test_utils::read_scan(&scan, engine_ref)?;
-    let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-    assert_eq!(total_rows, 6);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_create_table_with_data_uses_relative_paths() -> Result<(), Box<dyn std::error::Error>>
-{
-    let schema = get_simple_schema();
-    let (_tmp_dir, table_path, engine) = test_table_setup()?;
-    let table_url = Url::from_directory_path(&table_path).unwrap();
-
-    let mut txn = create_table_txn(table_url.as_str(), schema.clone(), "test/1.0")
-        .build(engine.as_ref(), Box::new(FileSystemCommitter::new()))?;
-    let write_context = txn.unpartitioned_write_context()?;
-    let add_meta = engine
-        .write_parquet(
-            &ArrowEngineData::new(simple_id_batch(&schema, vec![10, 20])),
-            &write_context,
-        )
-        .await?;
-    txn.add_files(add_meta);
-    let committed = txn.commit(engine.as_ref())?.unwrap_committed();
-    let snapshot = committed.post_commit_snapshot().unwrap().clone();
-
-    let add_infos = read_add_infos(&snapshot, engine.as_ref())?;
-    assert_eq!(add_infos.len(), 1);
-    let path = &add_infos[0].path;
-    assert!(
-        !path.contains("://"),
-        "should produce relative paths, got: {path}"
-    );
-
-    // Verify data is readable
-    let scan = snapshot.scan_builder().build()?;
-    let engine_ref: Arc<dyn Engine> = engine;
-    let batches = test_utils::read_scan(&scan, engine_ref)?;
-    assert_eq!(batches.len(), 1);
-    assert_eq!(batches[0].num_rows(), 2);
-
-    Ok(())
-}
-
-/// Builds a RecordBatch with schema (id: long, tags: array<string>, props: map<string, long>,
-/// v: variant). Each row gets one entry in tags, one entry in props, and a simple integer variant.
-/// If `nulls` is provided, it sets the null buffer for tags, props, and variant columns.
-fn complex_type_batch(schema: &StructType, ids: &[i64], nulls: Option<&[bool]>) -> RecordBatch {
-    let arrow_schema: ArrowSchema = schema.try_into_arrow().unwrap();
-    let arrow_schema = Arc::new(arrow_schema);
-    let n = ids.len();
-
-    let id_array = Int64Array::from(ids.to_vec());
-
-    // tags: one string element per row
-    let tag_values = StringArray::from((0..n).map(|i| format!("t{i}")).collect::<Vec<_>>());
-    let tag_offsets = OffsetBuffer::new((0..=n).map(|i| i as i32).collect::<Vec<_>>().into());
-    let list_field = match arrow_schema.field_with_name("tags").unwrap().data_type() {
-        ArrowDataType::List(f) => f.clone(),
-        other => panic!("expected List, got {other:?}"),
-    };
-    let tag_array = ListArray::new(
-        list_field,
-        tag_offsets,
-        Arc::new(tag_values),
-        nulls.map(NullBuffer::from),
-    );
-
-    // props: one {"k": id} entry per row
-    let map_keys = StringArray::from(vec!["k"; n]);
-    let map_vals = Int64Array::from(ids.to_vec());
-    let (map_entries_field, map_sorted) =
-        match arrow_schema.field_with_name("props").unwrap().data_type() {
-            ArrowDataType::Map(f, sorted) => (f.clone(), *sorted),
-            other => panic!("expected Map, got {other:?}"),
-        };
-    let entries_fields = match map_entries_field.data_type() {
-        ArrowDataType::Struct(f) => f.clone(),
-        other => panic!("expected Struct, got {other:?}"),
-    };
-    let entries = StructArray::new(
-        entries_fields,
-        vec![
-            Arc::new(map_keys) as ArrayRef,
-            Arc::new(map_vals) as ArrayRef,
-        ],
-        None,
-    );
-    let map_offsets = OffsetBuffer::new((0..=n).map(|i| i as i32).collect::<Vec<_>>().into());
-    let map_array = MapArray::new(
-        map_entries_field,
-        map_offsets,
-        entries,
-        nulls.map(NullBuffer::from),
-        map_sorted,
-    );
-
-    // v: simple integer variant per row (metadata=[0x01,0x00,0x00], value=[0x0C, low_byte])
-    let variant_meta = BinaryArray::from(
-        ids.iter()
-            .map(|_| Some(&[0x01u8, 0x00, 0x00][..]))
-            .collect::<Vec<_>>(),
-    );
-    let variant_val_data: Vec<[u8; 2]> = ids.iter().map(|&id| [0x0Cu8, id as u8]).collect();
-    let variant_val = BinaryArray::from(
-        variant_val_data
-            .iter()
-            .map(|v| Some(&v[..]))
-            .collect::<Vec<_>>(),
-    );
-    let variant_fields = match arrow_schema.field_with_name("v").unwrap().data_type() {
-        ArrowDataType::Struct(fields) => fields.clone(),
-        other => panic!("expected Struct, got {other:?}"),
-    };
-    let variant_array = StructArray::new(
-        variant_fields,
-        vec![
-            Arc::new(variant_meta) as ArrayRef,
-            Arc::new(variant_val) as ArrayRef,
-        ],
-        nulls.map(NullBuffer::from),
-    );
-
-    RecordBatch::try_new(
-        arrow_schema,
-        vec![
-            Arc::new(id_array) as ArrayRef,
-            Arc::new(tag_array) as ArrayRef,
-            Arc::new(map_array) as ArrayRef,
-            Arc::new(variant_array) as ArrayRef,
-        ],
-    )
-    .unwrap()
-}
-
-/// Verifies that writing a table with array, map, and variant columns produces nullCount
-/// statistics for those columns while excluding them from minValues/maxValues. Then writes a
-/// second file and scans with a predicate to verify data skipping works e2e. Parameterized
-/// over column mapping mode and whether a checkpoint is taken before the scan.
-#[rstest::rstest]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_write_stats_for_complex_type_columns(
-    #[values(
-        ColumnMappingMode::None,
-        ColumnMappingMode::Id,
-        ColumnMappingMode::Name
-    )]
-    cm_mode: ColumnMappingMode,
-    #[values(false, true)] use_checkpoint: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("id", DataType::LONG),
-        StructField::nullable(
-            "tags",
-            DataType::Array(Box::new(ArrayType::new(DataType::STRING, true))),
-        ),
-        StructField::nullable(
-            "props",
-            DataType::Map(Box::new(MapType::new(
-                DataType::STRING,
-                DataType::LONG,
-                true,
-            ))),
-        ),
-        StructField::nullable("v", DataType::unshredded_variant()),
-    ])?);
-
-    let mode_str = match cm_mode {
-        ColumnMappingMode::None => "none",
-        ColumnMappingMode::Id => "id",
-        ColumnMappingMode::Name => "name",
-    };
-    let (_tmp_dir, table_path, engine) = test_table_setup_mt()?;
-    let snapshot = create_table_and_load_snapshot(
-        &table_path,
-        schema.clone(),
-        engine.as_ref(),
-        &[("delta.columnMapping.mode", mode_str)],
-    )?;
-    let table_url = Url::from_directory_path(&table_path).unwrap();
-
-    // Resolve physical column names for stats JSON verification
-    let cm = snapshot
-        .table_properties()
-        .column_mapping_mode
-        .unwrap_or(ColumnMappingMode::None);
-    let physical_name = |logical: &str| -> String {
-        get_any_level_column_physical_name(
-            snapshot.schema().as_ref(),
-            &ColumnName::new([logical]),
-            cm,
-        )
-        .unwrap()
-        .into_inner()
-        .into_iter()
-        .next()
-        .unwrap()
-    };
-    let id_phys = physical_name("id");
-    let tags_phys = physical_name("tags");
-    let props_phys = physical_name("props");
-    let v_phys = physical_name("v");
-
-    // Batch 1: ids [1,2,3] with one null per complex column
-    let batch1 = complex_type_batch(&schema, &[1, 2, 3], Some(&[true, false, true]));
-    let _snapshot =
-        write_batch_to_table(&snapshot, engine.as_ref(), batch1, HashMap::new()).await?;
-
-    // Read the commit and verify stats use correct (physical) column names
-    let add_actions = read_actions_from_commit(&table_url, 1, "add")?;
-    assert_eq!(add_actions.len(), 1);
-
-    let stats: serde_json::Value = serde_json::from_str(
-        add_actions[0]
-            .get("stats")
-            .and_then(|s| s.as_str())
-            .expect("add action should have stats"),
-    )?;
-
-    assert_eq!(stats["numRecords"], 3);
-
-    // nullCount should be present for all columns including array, map, and variant
-    assert_eq!(stats["nullCount"][&id_phys], 0);
-    assert_eq!(stats["nullCount"][&tags_phys], 1);
-    assert_eq!(stats["nullCount"][&props_phys], 1);
-    assert_eq!(stats["nullCount"][&v_phys], 1);
-
-    // minValues/maxValues should have id but NOT complex types
-    assert!(stats["minValues"][&id_phys].is_number());
-    assert!(stats["maxValues"][&id_phys].is_number());
-    for col in [&tags_phys, &props_phys, &v_phys] {
-        assert!(
-            stats["minValues"].get(col).is_none(),
-            "minValues should not contain {col}"
-        );
-        assert!(
-            stats["maxValues"].get(col).is_none(),
-            "maxValues should not contain {col}"
-        );
-    }
-
-    // Batch 2: ids [10,11,12] with no nulls, written to a separate file
-    let batch2 = complex_type_batch(&schema, &[10, 11, 12], None);
-    let snapshot2 = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let snapshot2 =
-        write_batch_to_table(&snapshot2, engine.as_ref(), batch2, HashMap::new()).await?;
-
-    // Optionally checkpoint to verify stats survive the checkpoint round-trip
-    let scan_snapshot = if use_checkpoint {
-        snapshot2.checkpoint(engine.as_ref())?;
-        Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?
-    } else {
-        snapshot2
-    };
-
-    // Scan with predicate id > 5. File 1 has id in [1,3], file 2 has id in [10,12].
-    // Data skipping should skip file 1 entirely and return only file 2's rows.
-    let scan = scan_snapshot
-        .scan_builder()
-        .with_predicate(Arc::new(Pred::gt(column_expr!("id"), Expr::literal(5_i64))))
-        .build()?;
-    let batches: Vec<RecordBatch> = scan
-        .execute(engine.clone())?
-        .map(|r| {
-            let data = r.unwrap();
-            ArrowEngineData::try_from_engine_data(data)
-                .unwrap()
-                .record_batch()
-                .clone()
-        })
-        .collect();
-
-    let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-    assert_eq!(
-        total_rows, 3,
-        "predicate id > 5 should return only the second file's 3 rows"
-    );
-
-    let result_schema = batches[0].schema();
-    let combined = delta_kernel::arrow::compute::concat_batches(&result_schema, &batches)?;
-    let ids: Vec<i64> = combined
-        .column_by_name("id")
-        .unwrap()
-        .as_any()
-        .downcast_ref::<Int64Array>()
-        .unwrap()
-        .values()
-        .iter()
-        .copied()
-        .collect();
-    assert_eq!(ids, vec![10, 11, 12]);
-
-    Ok(())
-}
-
-/// Verifies that complex types in a nested schema count against `dataSkippingNumIndexedCols`.
-/// Schema: (id: long, data: struct<name: string, tags: array<string>, props: map<string, long>>).
-/// With numIndexedCols=3, the first 3 leaf columns (id, data.name, data.tags) get stats. The
-/// 4th leaf (data.props) is excluded because the limit is reached.
-#[tokio::test]
-async fn test_write_stats_nested_complex_types_respect_column_limit(
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    let schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("id", DataType::LONG),
-        StructField::nullable(
-            "data",
-            DataType::try_struct_type(vec![
-                StructField::nullable("name", DataType::STRING),
-                StructField::nullable(
-                    "tags",
-                    DataType::Array(Box::new(ArrayType::new(DataType::STRING, true))),
-                ),
-                StructField::nullable(
-                    "props",
-                    DataType::Map(Box::new(MapType::new(
-                        DataType::STRING,
-                        DataType::LONG,
-                        true,
-                    ))),
-                ),
-            ])?,
-        ),
-    ])?);
-
-    let (_tmp_dir, table_path, engine) = test_table_setup()?;
-    let table_url = Url::from_directory_path(&table_path).unwrap();
-    let snapshot =
-        create_table_and_load_snapshot(&table_path, schema.clone(), engine.as_ref(), &[])?;
-    let snapshot = set_table_properties(
-        &table_path,
-        &table_url,
-        engine.as_ref(),
-        snapshot.version(),
-        &[("delta.dataSkippingNumIndexedCols", "3")],
-    )?;
-
-    // Build a batch with 3 rows
-    let arrow_schema: ArrowSchema = schema.as_ref().try_into_arrow()?;
-    let arrow_schema = Arc::new(arrow_schema);
-
-    let id_array = Int64Array::from(vec![1, 2, 3]);
-    let name_array = StringArray::from(vec!["a", "b", "c"]);
-
-    // tags: [["x"], null, ["y"]]
-    let data_field = arrow_schema.field_with_name("data").unwrap();
-    let data_struct_fields = match data_field.data_type() {
-        ArrowDataType::Struct(f) => f,
-        other => panic!("expected Struct, got {other:?}"),
-    };
-    let tags_field = &data_struct_fields[1];
-    let list_field = match tags_field.data_type() {
-        ArrowDataType::List(f) => f.clone(),
-        other => panic!("expected List, got {other:?}"),
-    };
-    let tag_values = StringArray::from(vec!["x", "y"]);
-    let tag_offsets = OffsetBuffer::new(vec![0, 1, 1, 2].into());
-    let tag_array = ListArray::new(
-        list_field,
-        tag_offsets,
-        Arc::new(tag_values),
-        Some(NullBuffer::from_iter([true, false, true])),
-    );
-
-    // props: [{"k": 1}, {"k": 2}, null]
-    let props_field = &data_struct_fields[2];
-    let (map_entries_field, map_sorted) = match props_field.data_type() {
-        ArrowDataType::Map(f, sorted) => (f.clone(), *sorted),
-        other => panic!("expected Map, got {other:?}"),
-    };
-    let entries_fields = match map_entries_field.data_type() {
-        ArrowDataType::Struct(f) => f.clone(),
-        other => panic!("expected Struct, got {other:?}"),
-    };
-    let map_keys = StringArray::from(vec!["k", "k"]);
-    let map_vals = Int64Array::from(vec![1i64, 2]);
-    let entries = StructArray::new(
-        entries_fields,
-        vec![
-            Arc::new(map_keys) as ArrayRef,
-            Arc::new(map_vals) as ArrayRef,
-        ],
-        None,
-    );
-    let map_offsets = OffsetBuffer::new(vec![0, 1, 2, 2].into());
-    let map_array = MapArray::new(
-        map_entries_field,
-        map_offsets,
-        entries,
-        Some(NullBuffer::from_iter([true, true, false])),
-        map_sorted,
-    );
-
-    // Assemble the data struct
-    let data_array = StructArray::new(
-        data_struct_fields.clone(),
-        vec![
-            Arc::new(name_array) as ArrayRef,
-            Arc::new(tag_array) as ArrayRef,
-            Arc::new(map_array) as ArrayRef,
-        ],
-        None,
-    );
-
-    let batch = RecordBatch::try_new(
-        arrow_schema,
-        vec![
-            Arc::new(id_array) as ArrayRef,
-            Arc::new(data_array) as ArrayRef,
-        ],
-    )?;
-
-    let _snapshot = write_batch_to_table(&snapshot, engine.as_ref(), batch, HashMap::new()).await?;
-
-    // Version 0=create, 1=set properties, 2=data write
-    let add_actions = read_actions_from_commit(&table_url, 2, "add")?;
-    let stats: serde_json::Value = serde_json::from_str(
-        add_actions[0]
-            .get("stats")
-            .and_then(|s| s.as_str())
-            .expect("add action should have stats"),
-    )?;
-
-    assert_eq!(stats["numRecords"], 3);
-
-    // First 3 leaves: id, data.name, data.tags all get nullCount
-    assert_eq!(stats["nullCount"]["id"], 0);
-    assert_eq!(stats["nullCount"]["data"]["name"], 0);
-    assert_eq!(stats["nullCount"]["data"]["tags"], 1);
-
-    // 4th leaf data.props is excluded by the column limit
-    assert!(
-        stats["nullCount"]["data"].get("props").is_none(),
-        "props should be excluded by numIndexedCols=3"
-    );
-
-    // id and data.name get min/max; data.tags does not (complex type)
-    assert!(stats["minValues"]["id"].is_number());
-    assert!(stats["minValues"]["data"]["name"].is_string());
-    assert!(stats["minValues"]["data"].get("tags").is_none());
-
-    Ok(())
-}
